Network.AWS.CloudSearch

Description

Amazon CloudSearch is a fully-managed service in the AWS Cloud that makes it simple and cost-effective to set up, manage, and scale a search solution for your website or application. Amazon CloudSearch supports 34 languages and popular search features such as highlighting, autocomplete, and geospatial search.

Documentation

module Network.AWS.CloudSearch.BuildSuggesters

module Network.AWS.CloudSearch.CreateDomain

module Network.AWS.CloudSearch.DefineAnalysisScheme

module Network.AWS.CloudSearch.DefineExpression

module Network.AWS.CloudSearch.DefineIndexField

module Network.AWS.CloudSearch.DefineSuggester

module Network.AWS.CloudSearch.DeleteAnalysisScheme

module Network.AWS.CloudSearch.DeleteDomain

module Network.AWS.CloudSearch.DeleteExpression

module Network.AWS.CloudSearch.DeleteIndexField

module Network.AWS.CloudSearch.DeleteSuggester

module Network.AWS.CloudSearch.DescribeAnalysisSchemes

module Network.AWS.CloudSearch.DescribeAvailabilityOptions

module Network.AWS.CloudSearch.DescribeDomains

module Network.AWS.CloudSearch.DescribeExpressions

module Network.AWS.CloudSearch.DescribeIndexFields

module Network.AWS.CloudSearch.DescribeScalingParameters

module Network.AWS.CloudSearch.DescribeServiceAccessPolicies

module Network.AWS.CloudSearch.DescribeSuggesters

module Network.AWS.CloudSearch.IndexDocuments

module Network.AWS.CloudSearch.ListDomainNames

module Network.AWS.CloudSearch.Types

module Network.AWS.CloudSearch.UpdateAvailabilityOptions

module Network.AWS.CloudSearch.UpdateScalingParameters

module Network.AWS.CloudSearch.UpdateServiceAccessPolicies