Network.AWS.CloudSearch.BuildSuggesters

Description

Indexes the search suggestions. For more information, see Configuring Suggesters in the Amazon CloudSearch Developer Guide .

Creating a Request

buildSuggesters

Arguments

:: Text
bsDomainName

-> BuildSuggesters

Creates a value of BuildSuggesters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bsDomainName - Undocumented member.

data BuildSuggesters

Container for the parameters to the BuildSuggester operation. Specifies the name of the domain you want to update.

See: buildSuggesters smart constructor.

Request Lenses

bsDomainName :: Lens' BuildSuggesters Text

Undocumented member.

Destructuring the Response

buildSuggestersResponse

Arguments

:: Int
bsrsResponseStatus

-> BuildSuggestersResponse

Creates a value of BuildSuggestersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bsrsFieldNames - Undocumented member.

bsrsResponseStatus - -- | The response status code.

data BuildSuggestersResponse

The result of a BuildSuggester request. Contains a list of the fields used for suggestions.

See: buildSuggestersResponse smart constructor.

Response Lenses

bsrsFieldNames :: Lens' BuildSuggestersResponse [Text]

Undocumented member.

bsrsResponseStatus :: Lens' BuildSuggestersResponse Int

- | The response status code.