Network.AWS.CloudSearch.CreateDomain

Description

Creates a new search domain. For more information, see Creating a Search Domain in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_CreateDomain.html

Request

data CreateDomain

Request constructor

createDomain

Arguments

:: Text
cdDomainName

-> CreateDomain

CreateDomain constructor.

The fields accessible through corresponding lenses are:

cdDomainName :: Text

Request lenses

cdDomainName :: Lens' CreateDomain Text

A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

Response

data CreateDomainResponse

Response constructor

createDomainResponse :: CreateDomainResponse

CreateDomainResponse constructor.

The fields accessible through corresponding lenses are:

cdrDomainStatus :: Maybe DomainStatus

Response lenses

cdrDomainStatus :: Lens' CreateDomainResponse (Maybe DomainStatus)