Network.AWS.CloudSearch.DefineAnalysisScheme

Description

Configures an analysis scheme that can be applied to a text or text-array field to define language-specific text processing options. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Request

data DefineAnalysisScheme

Container for the parameters to the DefineAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme configuration.

Request constructor

defineAnalysisScheme

Arguments

:: Text
dasDomainName

-> AnalysisScheme
dasAnalysisScheme

-> DefineAnalysisScheme

Smart constructor for the minimum required parameters to construct a valid DefineAnalysisScheme request.

The fields accessible through corresponding lenses are:

DomainName :: Text

AnalysisScheme :: AnalysisScheme

Request lenses

dasDomainName :: Lens' DefineAnalysisScheme Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dasAnalysisScheme :: Lens' DefineAnalysisScheme AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

Response

data DefineAnalysisSchemeResponse

The result of a DefineAnalysisScheme request. Contains the status of the newly-configured analysis scheme.

Response constructor

defineAnalysisSchemeResponse

Arguments

:: AnalysisSchemeStatus
dasrAnalysisScheme

-> DefineAnalysisSchemeResponse

Smart constructor for the minimum required parameters to construct a valid DefineAnalysisSchemeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AnalysisScheme :: AnalysisSchemeStatus

Response lenses

dasrAnalysisScheme :: Lens' DefineAnalysisSchemeResponse AnalysisSchemeStatus

The status and configuration of an AnalysisScheme.