Network.AWS.CloudSearch.DefineExpression

Description

Configures an 'Expression for the search domain. Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DefineExpression.html

Request

data DefineExpression

Request constructor

defineExpression

Arguments

:: Text
de1DomainName

-> Expression
de1Expression

-> DefineExpression

DefineExpression constructor.

The fields accessible through corresponding lenses are:

de1DomainName :: Text

de1Expression :: Expression

Request lenses

de1DomainName :: Lens' DefineExpression Text

de1Expression :: Lens' DefineExpression Expression

Response

data DefineExpressionResponse

Response constructor

defineExpressionResponse

Arguments

:: ExpressionStatus
derExpression

-> DefineExpressionResponse

DefineExpressionResponse constructor.

The fields accessible through corresponding lenses are:

derExpression :: ExpressionStatus

Response lenses

derExpression :: Lens' DefineExpressionResponse ExpressionStatus