Network.AWS.CloudSearch.DefineIndexField

Description

Configures an 'IndexField for the search domain. Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the 'IndexFieldType. If the field exists, the new configuration replaces the old one. For more information, see Configuring Index Fields in the Amazon CloudSearch DeveloperGuide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DefineIndexField.html

Request

data DefineIndexField

Request constructor

defineIndexField

Arguments

:: Text
dif2DomainName

-> IndexField
dif2IndexField

-> DefineIndexField

DefineIndexField constructor.

The fields accessible through corresponding lenses are:

dif2DomainName :: Text

dif2IndexField :: IndexField

Request lenses

dif2DomainName :: Lens' DefineIndexField Text

dif2IndexField :: Lens' DefineIndexField IndexField

The index field and field options you want to configure.

Response

data DefineIndexFieldResponse

Response constructor

defineIndexFieldResponse

Arguments

:: IndexFieldStatus
difr1IndexField

-> DefineIndexFieldResponse

DefineIndexFieldResponse constructor.

The fields accessible through corresponding lenses are:

difr1IndexField :: IndexFieldStatus

Response lenses

difr1IndexField :: Lens' DefineIndexFieldResponse IndexFieldStatus