Network.AWS.CloudSearch.DefineSuggester

Description

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Request

data DefineSuggester

Container for the parameters to the DefineSuggester operation. Specifies the name of the domain you want to update and the suggester configuration.

Request constructor

defineSuggester

Arguments

:: Text
ds1DomainName

-> Suggester
ds1Suggester

-> DefineSuggester

Smart constructor for the minimum required parameters to construct a valid DefineSuggester request.

The fields accessible through corresponding lenses are:

DomainName :: Text

Suggester :: Suggester

Request lenses

ds1DomainName :: Lens' DefineSuggester Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

ds1Suggester :: Lens' DefineSuggester Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

Response

data DefineSuggesterResponse

The result of a DefineSuggester request. Contains the status of the newly-configured suggester.

Response constructor

defineSuggesterResponse

Arguments

:: SuggesterStatus
dsrSuggester

-> DefineSuggesterResponse

Smart constructor for the minimum required parameters to construct a valid DefineSuggesterResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Suggester :: SuggesterStatus

Response lenses

dsrSuggester :: Lens' DefineSuggesterResponse SuggesterStatus

The value of a Suggester and its current status.