Network.AWS.CloudSearch.DeleteAnalysisScheme

Description

Deletes an analysis scheme. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DeleteAnalysisScheme.html

Request

data DeleteAnalysisScheme

Request constructor

deleteAnalysisScheme

Arguments

:: Text
dasDomainName

-> Text
dasAnalysisSchemeName

-> DeleteAnalysisScheme

DeleteAnalysisScheme constructor.

The fields accessible through corresponding lenses are:

dasAnalysisSchemeName :: Text

dasDomainName :: Text

Request lenses

dasAnalysisSchemeName :: Lens' DeleteAnalysisScheme Text

The name of the analysis scheme you want to delete.

dasDomainName :: Lens' DeleteAnalysisScheme Text

Response

data DeleteAnalysisSchemeResponse

Response constructor

deleteAnalysisSchemeResponse

Arguments

:: AnalysisSchemeStatus
dasrAnalysisScheme

-> DeleteAnalysisSchemeResponse

DeleteAnalysisSchemeResponse constructor.

The fields accessible through corresponding lenses are:

dasrAnalysisScheme :: AnalysisSchemeStatus

Response lenses

dasrAnalysisScheme :: Lens' DeleteAnalysisSchemeResponse AnalysisSchemeStatus

The status of the analysis scheme being deleted.