Network.AWS.CloudSearch.DeleteDomain

Description

Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see Deleting a Search Domain in the Amazon CloudSearch Developer Guide .

Creating a Request

deleteDomain

Arguments

:: Text
dddDomainName

-> DeleteDomain

Creates a value of DeleteDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dddDomainName - The name of the domain you want to permanently delete.

data DeleteDomain

Container for the parameters to the DeleteDomain operation. Specifies the name of the domain you want to delete.

See: deleteDomain smart constructor.

Request Lenses

dddDomainName :: Lens' DeleteDomain Text

The name of the domain you want to permanently delete.

Destructuring the Response

deleteDomainResponse

Arguments

:: Int
ddrsResponseStatus

-> DeleteDomainResponse

Creates a value of DeleteDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddrsDomainStatus - Undocumented member.

ddrsResponseStatus - -- | The response status code.

data DeleteDomainResponse

The result of a DeleteDomain request. Contains the status of a newly deleted domain, or no status if the domain has already been completely deleted.

See: deleteDomainResponse smart constructor.

Response Lenses

ddrsDomainStatus :: Lens' DeleteDomainResponse (Maybe DomainStatus)

Undocumented member.

ddrsResponseStatus :: Lens' DeleteDomainResponse Int

- | The response status code.