Network.AWS.CloudSearch.DeleteExpression

Description

Removes an Expression from the search domain. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Request

data DeleteExpression

Container for the parameters to the DeleteExpression operation. Specifies the name of the domain you want to update and the name of the expression you want to delete.

Request constructor

deleteExpression

Arguments

:: Text
de1DomainName

-> Text
de1ExpressionName

-> DeleteExpression

Smart constructor for the minimum required parameters to construct a valid DeleteExpression request.

The fields accessible through corresponding lenses are:

DomainName :: Text

ExpressionName :: Text

Request lenses

de1DomainName :: Lens' DeleteExpression Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

de1ExpressionName :: Lens' DeleteExpression Text

The name of the Expression to delete.

Response

data DeleteExpressionResponse

The result of a DeleteExpression request. Specifies the expression being deleted.

Response constructor

deleteExpressionResponse

Arguments

:: ExpressionStatus
derrExpression

-> DeleteExpressionResponse

Smart constructor for the minimum required parameters to construct a valid DeleteExpressionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Expression :: ExpressionStatus

Response lenses

derrExpression :: Lens' DeleteExpressionResponse ExpressionStatus

The status of the expression being deleted.