Network.AWS.CloudSearch.DeleteIndexField

Description

Removes an IndexField from the search domain. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Request

data DeleteIndexField

Container for the parameters to the DeleteIndexField operation. Specifies the name of the domain you want to update and the name of the index field you want to delete.

Request constructor

deleteIndexField

Arguments

:: Text
dif1DomainName

-> Text
dif1IndexFieldName

-> DeleteIndexField

Smart constructor for the minimum required parameters to construct a valid DeleteIndexField request.

The fields accessible through corresponding lenses are:

DomainName :: Text

IndexFieldName :: Text

Request lenses

dif1DomainName :: Lens' DeleteIndexField Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dif1IndexFieldName :: Lens' DeleteIndexField Text

The name of the index field your want to remove from the domain's indexing options.

Response

data DeleteIndexFieldResponse

The result of a DeleteIndexField request.

Response constructor

deleteIndexFieldResponse

Arguments

:: IndexFieldStatus
difrrIndexField

-> DeleteIndexFieldResponse

Smart constructor for the minimum required parameters to construct a valid DeleteIndexFieldResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IndexField :: IndexFieldStatus

Response lenses

difrrIndexField :: Lens' DeleteIndexFieldResponse IndexFieldStatus

The status of the index field being deleted.