Network.AWS.CloudSearch.DeleteSuggester

Description

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DeleteSuggester.html

Request

data DeleteSuggester

Request constructor

deleteSuggester

Arguments

:: Text
ds3DomainName

-> Text
ds3SuggesterName

-> DeleteSuggester

DeleteSuggester constructor.

The fields accessible through corresponding lenses are:

ds3DomainName :: Text

ds3SuggesterName :: Text

Request lenses

ds3DomainName :: Lens' DeleteSuggester Text

ds3SuggesterName :: Lens' DeleteSuggester Text

Specifies the name of the suggester you want to delete.

Response

data DeleteSuggesterResponse

Response constructor

deleteSuggesterResponse

Arguments

:: SuggesterStatus
dsr1Suggester

-> DeleteSuggesterResponse

DeleteSuggesterResponse constructor.

The fields accessible through corresponding lenses are:

dsr1Suggester :: SuggesterStatus

Response lenses

dsr1Suggester :: Lens' DeleteSuggesterResponse SuggesterStatus

The status of the suggester being deleted.