Network.AWS.CloudSearch.DescribeAnalysisSchemes

Description

Gets the analysis schemes configured for a domain. An analysis scheme defines language-specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Analysis Schemes in the Amazon CloudSearchDeveloper Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeAnalysisSchemes.html

Request

data DescribeAnalysisSchemes

Request constructor

describeAnalysisSchemes

Arguments

:: Text
das1DomainName

-> DescribeAnalysisSchemes

DescribeAnalysisSchemes constructor.

The fields accessible through corresponding lenses are:

das1AnalysisSchemeNames :: [Text]

das1Deployed :: Maybe Bool

das1DomainName :: Text

Request lenses

das1AnalysisSchemeNames :: Lens' DescribeAnalysisSchemes [Text]

The analysis schemes you want to describe.

das1Deployed :: Lens' DescribeAnalysisSchemes (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

das1DomainName :: Lens' DescribeAnalysisSchemes Text

The name of the domain you want to describe.

Response

data DescribeAnalysisSchemesResponse

Response constructor

describeAnalysisSchemesResponse :: DescribeAnalysisSchemesResponse

DescribeAnalysisSchemesResponse constructor.

The fields accessible through corresponding lenses are:

dasrAnalysisSchemes :: [AnalysisSchemeStatus]

Response lenses

dasrAnalysisSchemes :: Lens' DescribeAnalysisSchemesResponse [AnalysisSchemeStatus]

The analysis scheme descriptions.