Network.AWS.CloudSearch.DescribeAvailabilityOptions

Description

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

Request

data DescribeAvailabilityOptions

Container for the parameters to the DescribeAvailabilityOptions operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

Request constructor

mkDescribeAvailabilityOptions

Arguments

:: Text
dao2DomainName

-> DescribeAvailabilityOptions

Smart constructor for the minimum required parameters to construct a valid DescribeAvailabilityOptions request.

The fields accessible through corresponding lenses are:

DomainName :: Text

Deployed :: Maybe Bool

Request lenses

dao2DomainName :: Lens' DescribeAvailabilityOptions Text

The name of the domain you want to describe.

dao2Deployed :: Lens' DescribeAvailabilityOptions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Response

data DescribeAvailabilityOptionsResponse

The result of a DescribeAvailabilityOptions request. Indicates whether or not the Multi-AZ option is enabled for the domain specified in the request.

Response constructor

mkDescribeAvailabilityOptionsResponse :: DescribeAvailabilityOptionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAvailabilityOptionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AvailabilityOptions :: Maybe AvailabilityOptionsStatus

Response lenses

daorrAvailabilityOptions :: Lens' DescribeAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.