Network.AWS.CloudSearch.DescribeDomains

Description

Gets information about the search domains owned by this account. Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: 'q=matchall&amp;q.parser=structured&amp;size=0'. For more information, see Getting Information about a Search Domain in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeDomains.html

Request

data DescribeDomains

Request constructor

describeDomains :: DescribeDomains

DescribeDomains constructor.

The fields accessible through corresponding lenses are:

ddDomainNames :: [Text]

Request lenses

ddDomainNames :: Lens' DescribeDomains [Text]

The names of the domains you want to include in the response.

Response

data DescribeDomainsResponse

Response constructor

describeDomainsResponse :: DescribeDomainsResponse

DescribeDomainsResponse constructor.

The fields accessible through corresponding lenses are:

ddrDomainStatusList :: [DomainStatus]

Response lenses

ddrDomainStatusList :: Lens' DescribeDomainsResponse [DomainStatus]