Network.AWS.CloudSearch.DescribeExpressions

Description

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch DeveloperGuide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeExpressions.html

Request

data DescribeExpressions

Request constructor

describeExpressions

Arguments

:: Text
deDomainName

-> DescribeExpressions

DescribeExpressions constructor.

The fields accessible through corresponding lenses are:

deDeployed :: Maybe Bool

deDomainName :: Text

deExpressionNames :: [Text]

Request lenses

deDeployed :: Lens' DescribeExpressions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

deDomainName :: Lens' DescribeExpressions Text

The name of the domain you want to describe.

deExpressionNames :: Lens' DescribeExpressions [Text]

Limits the 'DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

Response

data DescribeExpressionsResponse

Response constructor

describeExpressionsResponse :: DescribeExpressionsResponse

DescribeExpressionsResponse constructor.

The fields accessible through corresponding lenses are:

derExpressions :: [ExpressionStatus]

Response lenses

derExpressions :: Lens' DescribeExpressionsResponse [ExpressionStatus]

The expressions configured for the domain.