Network.AWS.CloudSearch.DescribeIndexFields

Description

Gets information about the index fields configured for the search domain. Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Domain Information in the Amazon CloudSearchDeveloper Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeIndexFields.html

Request

data DescribeIndexFields

Request constructor

describeIndexFields

Arguments

:: Text
difDomainName

-> DescribeIndexFields

DescribeIndexFields constructor.

The fields accessible through corresponding lenses are:

difDeployed :: Maybe Bool

difDomainName :: Text

difFieldNames :: [Text]

Request lenses

difDeployed :: Lens' DescribeIndexFields (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

difDomainName :: Lens' DescribeIndexFields Text

The name of the domain you want to describe.

difFieldNames :: Lens' DescribeIndexFields [Text]

A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

Response

data DescribeIndexFieldsResponse

Response constructor

describeIndexFieldsResponse :: DescribeIndexFieldsResponse

DescribeIndexFieldsResponse constructor.

The fields accessible through corresponding lenses are:

difrIndexFields :: [IndexFieldStatus]

Response lenses

difrIndexFields :: Lens' DescribeIndexFieldsResponse [IndexFieldStatus]

The index fields configured for the domain.