Network.AWS.CloudSearch.DescribeScalingParameters

Description

Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see Configuring Scaling Options in the AmazonCloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeScalingParameters.html

Request

data DescribeScalingParameters

Request constructor

describeScalingParameters

Arguments

:: Text
dspDomainName

-> DescribeScalingParameters

DescribeScalingParameters constructor.

The fields accessible through corresponding lenses are:

dspDomainName :: Text

Request lenses

dspDomainName :: Lens' DescribeScalingParameters Text

Response

data DescribeScalingParametersResponse

Response constructor

describeScalingParametersResponse

Arguments

:: ScalingParametersStatus
dsprScalingParameters

-> DescribeScalingParametersResponse

DescribeScalingParametersResponse constructor.

The fields accessible through corresponding lenses are:

dsprScalingParameters :: ScalingParametersStatus

Response lenses

dsprScalingParameters :: Lens' DescribeScalingParametersResponse ScalingParametersStatus