Network.AWS.CloudSearch.DescribeServiceAccessPolicies

Description

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

Request

data DescribeServiceAccessPolicies

Container for the parameters to the DescribeServiceAccessPolicies operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

Request constructor

mkDescribeServiceAccessPolicies

Arguments

:: Text
dsapDomainName

-> DescribeServiceAccessPolicies

Smart constructor for the minimum required parameters to construct a valid DescribeServiceAccessPolicies request.

The fields accessible through corresponding lenses are:

DomainName :: Text

Deployed :: Maybe Bool

Request lenses

dsapDomainName :: Lens' DescribeServiceAccessPolicies Text

The name of the domain you want to describe.

dsapDeployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Response

data DescribeServiceAccessPoliciesResponse

The result of a DescribeServiceAccessPolicies request.

Response constructor

mkDescribeServiceAccessPoliciesResponse

Arguments

:: AccessPoliciesStatus
dsaprAccessPolicies

-> DescribeServiceAccessPoliciesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeServiceAccessPoliciesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AccessPolicies :: AccessPoliciesStatus

Response lenses

dsaprAccessPolicies :: Lens' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain specified in the request.