Network.AWS.CloudSearch.DescribeSuggesters

Description

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide .

Creating a Request

describeSuggesters

Arguments

:: Text
dssDomainName

-> DescribeSuggesters

Creates a value of DescribeSuggesters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssDeployed - Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

dssSuggesterNames - The suggesters you want to describe.

dssDomainName - The name of the domain you want to describe.

data DescribeSuggesters

Container for the parameters to the DescribeSuggester operation. Specifies the name of the domain you want to describe. To restrict the response to particular suggesters, specify the names of the suggesters you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true .

See: describeSuggesters smart constructor.

Request Lenses

dssDeployed :: Lens' DescribeSuggesters (Maybe Bool)

Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

dssSuggesterNames :: Lens' DescribeSuggesters [Text]

The suggesters you want to describe.

dssDomainName :: Lens' DescribeSuggesters Text

The name of the domain you want to describe.

Destructuring the Response

describeSuggestersResponse

Arguments

:: Int
dssrsResponseStatus

-> DescribeSuggestersResponse

Creates a value of DescribeSuggestersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssrsResponseStatus - -- | The response status code.

dssrsSuggesters - The suggesters configured for the domain specified in the request.

data DescribeSuggestersResponse

The result of a DescribeSuggesters request.

See: describeSuggestersResponse smart constructor.

Response Lenses

dssrsResponseStatus :: Lens' DescribeSuggestersResponse Int

- | The response status code.

dssrsSuggesters :: Lens' DescribeSuggestersResponse [SuggesterStatus]

The suggesters configured for the domain specified in the request.