Network.AWS.CloudSearch.IndexDocuments

Description

Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_IndexDocuments.html

Request

data IndexDocuments

Request constructor

indexDocuments

Arguments

:: Text
idDomainName

-> IndexDocuments

IndexDocuments constructor.

The fields accessible through corresponding lenses are:

idDomainName :: Text

Request lenses

idDomainName :: Lens' IndexDocuments Text

Response

data IndexDocumentsResponse

Response constructor

indexDocumentsResponse :: IndexDocumentsResponse

IndexDocumentsResponse constructor.

The fields accessible through corresponding lenses are:

idrFieldNames :: [Text]

Response lenses

idrFieldNames :: Lens' IndexDocumentsResponse [Text]

The names of the fields that are currently being indexed.