Network.AWS.CloudSearch.Monadic

Description

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.CloudSearch with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.CloudSearch

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.CloudSearch.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Control.Applicative
import Network.AWS.CloudSearch.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

-- Or to void any additional parameters outside of those required using return ():
operationName w x $ return ()

BuildSuggesters

Indexes the search suggestions.

See: BuildSuggesters

buildSuggesters

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
bsDomainName

-> m BuildSuggestersResponse

buildSuggestersCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
bsDomainName

-> m (Either ServiceEr BuildSuggestersResponse)

CreateDomain

Creates a new search domain. For more information, see Creating a Search Domain in the Amazon CloudSearch Developer Guide.

See: CreateDomain

createDomain

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cdDomainName

-> m CreateDomainResponse

createDomainCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cdDomainName

-> m (Either ServiceEr CreateDomainResponse)

DefineAnalysisScheme

Configures an analysis scheme that can be applied to a text or text-array field to define language-specific text processing options. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

See: DefineAnalysisScheme

defineAnalysisScheme

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dasDomainName

-> AnalysisScheme
dasAnalysisScheme

-> m DefineAnalysisSchemeResponse

defineAnalysisSchemeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dasDomainName

-> AnalysisScheme
dasAnalysisScheme

-> m (Either ServiceEr DefineAnalysisSchemeResponse)

DefineExpression

Configures an Expression for the search domain. Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

See: DefineExpression

defineExpression

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
deDomainName

-> Expression
deExpression

-> m DefineExpressionResponse

defineExpressionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
deDomainName

-> Expression
deExpression

-> m (Either ServiceEr DefineExpressionResponse)

DefineIndexField

Configures an IndexField for the search domain. Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the IndexFieldType. If the field exists, the new configuration replaces the old one. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

See: DefineIndexField

defineIndexField

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
difDomainName

-> IndexField
difIndexField

-> m DefineIndexFieldResponse

defineIndexFieldCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
difDomainName

-> IndexField
difIndexField

-> m (Either ServiceEr DefineIndexFieldResponse)

DefineSuggester

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

See: DefineSuggester

defineSuggester

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ds1DomainName

-> Suggester
ds1Suggester

-> m DefineSuggesterResponse

defineSuggesterCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ds1DomainName

-> Suggester
ds1Suggester

-> m (Either ServiceEr DefineSuggesterResponse)

DeleteAnalysisScheme

Deletes an analysis scheme. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

See: DeleteAnalysisScheme

deleteAnalysisScheme

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
das1DomainName

-> Text
das1AnalysisSchemeName

-> m DeleteAnalysisSchemeResponse

deleteAnalysisSchemeCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
das1DomainName

-> Text
das1AnalysisSchemeName

-> m (Either ServiceEr DeleteAnalysisSchemeResponse)

DeleteDomain

Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see Deleting a Search Domain in the Amazon CloudSearch Developer Guide.

See: DeleteDomain

deleteDomain

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ddDomainName

-> m DeleteDomainResponse

deleteDomainCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ddDomainName

-> m (Either ServiceEr DeleteDomainResponse)

DeleteExpression

Removes an Expression from the search domain. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

See: DeleteExpression

deleteExpression

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
de1DomainName

-> Text
de1ExpressionName

-> m DeleteExpressionResponse

deleteExpressionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
de1DomainName

-> Text
de1ExpressionName

-> m (Either ServiceEr DeleteExpressionResponse)

DeleteIndexField

Removes an IndexField from the search domain. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

See: DeleteIndexField

deleteIndexField

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dif1DomainName

-> Text
dif1IndexFieldName

-> m DeleteIndexFieldResponse

deleteIndexFieldCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dif1DomainName

-> Text
dif1IndexFieldName

-> m (Either ServiceEr DeleteIndexFieldResponse)

DeleteSuggester

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

See: DeleteSuggester

deleteSuggester

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ds2DomainName

-> Text
ds2SuggesterName

-> m DeleteSuggesterResponse

deleteSuggesterCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ds2DomainName

-> Text
ds2SuggesterName

-> m (Either ServiceEr DeleteSuggesterResponse)

DescribeAnalysisSchemes

Gets the analysis schemes configured for a domain. An analysis scheme defines language-specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

See: DescribeAnalysisSchemes

describeAnalysisSchemes

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
das2DomainName

-> State DescribeAnalysisSchemes a

-> m DescribeAnalysisSchemesResponse

describeAnalysisSchemesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
das2DomainName

-> State DescribeAnalysisSchemes a

-> m (Either ServiceEr DescribeAnalysisSchemesResponse)

DescribeAvailabilityOptions

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

See: DescribeAvailabilityOptions

describeAvailabilityOptions

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dao2DomainName

-> State DescribeAvailabilityOptions a

-> m DescribeAvailabilityOptionsResponse

describeAvailabilityOptionsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dao2DomainName

-> State DescribeAvailabilityOptions a

-> m (Either ServiceEr DescribeAvailabilityOptionsResponse)

DescribeDomains

Gets information about the search domains owned by this account. Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: q=matchall&amp;q.parser=structured&amp;size=0. For more information, see Getting Information about a Search Domain in the Amazon CloudSearch Developer Guide.

See: DescribeDomains

describeDomains :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeDomains a -> m DescribeDomainsResponse

describeDomainsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeDomains a -> m (Either ServiceEr DescribeDomainsResponse)

DescribeExpressions

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

See: DescribeExpressions

describeExpressions

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
de2DomainName

-> State DescribeExpressions a

-> m DescribeExpressionsResponse

describeExpressionsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
de2DomainName

-> State DescribeExpressions a

-> m (Either ServiceEr DescribeExpressionsResponse)

DescribeIndexFields

Gets information about the index fields configured for the search domain. Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Domain Information in the Amazon CloudSearch Developer Guide.

See: DescribeIndexFields

describeIndexFields

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dif2DomainName

-> State DescribeIndexFields a

-> m DescribeIndexFieldsResponse

describeIndexFieldsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dif2DomainName

-> State DescribeIndexFields a

-> m (Either ServiceEr DescribeIndexFieldsResponse)

DescribeScalingParameters

Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

See: DescribeScalingParameters

describeScalingParameters

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dspDomainName

-> m DescribeScalingParametersResponse

describeScalingParametersCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dspDomainName

-> m (Either ServiceEr DescribeScalingParametersResponse)

DescribeServiceAccessPolicies

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

See: DescribeServiceAccessPolicies

describeServiceAccessPolicies

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dsapDomainName

-> State DescribeServiceAccessPolicies a

-> m DescribeServiceAccessPoliciesResponse

describeServiceAccessPoliciesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dsapDomainName

-> State DescribeServiceAccessPolicies a

-> m (Either ServiceEr DescribeServiceAccessPoliciesResponse)

DescribeSuggesters

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

See: DescribeSuggesters

describeSuggesters

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ds3DomainName

-> State DescribeSuggesters a

-> m DescribeSuggestersResponse

describeSuggestersCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ds3DomainName

-> State DescribeSuggesters a

-> m (Either ServiceEr DescribeSuggestersResponse)

IndexDocuments

Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments.

See: IndexDocuments

indexDocuments

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
idDomainName

-> m IndexDocumentsResponse

indexDocumentsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
idDomainName

-> m (Either ServiceEr IndexDocumentsResponse)

ListDomainNames

Lists all search domains owned by an account.

See: ListDomainNames

listDomainNames :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => m ListDomainNamesResponse

listDomainNamesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => m (Either ServiceEr ListDomainNamesResponse)

UpdateAvailabilityOptions

Configures the availability options for a domain. Enabling the Multi-AZ option expands an Amazon CloudSearch domain to an additional Availability Zone in the same Region to increase fault tolerance in the event of a service disruption. Changes to the Multi-AZ option can take about half an hour to become active. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

See: UpdateAvailabilityOptions

updateAvailabilityOptions

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uaoDomainName

-> Bool
uaoMultiAZ

-> m UpdateAvailabilityOptionsResponse

updateAvailabilityOptionsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uaoDomainName

-> Bool
uaoMultiAZ

-> m (Either ServiceEr UpdateAvailabilityOptionsResponse)

UpdateScalingParameters

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

See: UpdateScalingParameters

updateScalingParameters

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uspDomainName

-> ScalingParameters
uspScalingParameters

-> m UpdateScalingParametersResponse

updateScalingParametersCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uspDomainName

-> ScalingParameters
uspScalingParameters

-> m (Either ServiceEr UpdateScalingParametersResponse)

UpdateServiceAccessPolicies

Configures the access rules that control access to the domain's document and search endpoints. For more information, see Configuring Access for an Amazon CloudSearch Domain.

See: UpdateServiceAccessPolicies

updateServiceAccessPolicies

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
usapDomainName

-> Text
usapAccessPolicies

-> m UpdateServiceAccessPoliciesResponse

updateServiceAccessPoliciesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
usapDomainName

-> Text
usapAccessPolicies

-> m (Either ServiceEr UpdateServiceAccessPoliciesResponse)

Re-exported

module Network.AWS.CloudSearch

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()