Network.AWS.CloudSearch.Types

Description

Amazon CloudSearch is a fully-managed service in the AWS Cloud that makes it simple and cost-effective to set up, manage, and scale a search solution for your website or application. Amazon CloudSearch supports 34 languages and popular search features such as highlighting, autocomplete, and geospatial search.

Service

data CloudSearch

Supported version (2013-01-01) of the Amazon CloudSearch service.

Errors

data CloudSearchError

A sum type representing possible errors returned by the CloudSearch service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

BaseException
An error occurred while processing the request.

Fields

_beCode :: Maybe Text

_beMessage :: Maybe Text

CloudSearchClient HttpException

CloudSearchSerializer String

CloudSearchService String

DisabledOperationException
The request was rejected because it attempted an operation which is not enabled.

InternalException
An internal error occurred while processing the request. If this problem persists, report an issue from the Service Health Dashboard.

InvalidTypeException
The request was rejected because it specified an invalid type definition.

LimitExceededException
The request was rejected because a resource limit has already been met.

ResourceNotFoundException
The request was rejected because it attempted to reference a resource that does not exist.

_BaseException :: Prism' CloudSearchError (Maybe Text, Maybe Text)

An error occurred while processing the request.

See: BaseException

_CloudSearchClient :: Prism' CloudSearchError HttpException

See: CloudSearchClient

_CloudSearchSerializer :: Prism' CloudSearchError String

See: CloudSearchSerializer

_CloudSearchService :: Prism' CloudSearchError String

See: CloudSearchService

_DisabledOperationException :: Prism' CloudSearchError ()

The request was rejected because it attempted an operation which is not enabled.

See: DisabledOperationException

_InternalException :: Prism' CloudSearchError ()

An internal error occurred while processing the request. If this problem persists, report an issue from the Service Health Dashboard.

See: InternalException

_InvalidTypeException :: Prism' CloudSearchError ()

The request was rejected because it specified an invalid type definition.

See: InvalidTypeException

_LimitExceededException :: Prism' CloudSearchError ()

The request was rejected because a resource limit has already been met.

See: LimitExceededException

_ResourceNotFoundException :: Prism' CloudSearchError ()

The request was rejected because it attempted to reference a resource that does not exist.

See: ResourceNotFoundException

XML

xmlOptions :: Tagged a XMLOptions

AlgorithmicStemming

data AlgorithmicStemming

Constructors

AlgorithmicStemmingFull
full

AlgorithmicStemmingLight
light

AlgorithmicStemmingMinimal
minimal

AlgorithmicStemmingNone
none

AnalysisSchemeLanguage

data AnalysisSchemeLanguage

Constructors

AnalysisSchemeLanguageAr
ar

AnalysisSchemeLanguageBg
bg

AnalysisSchemeLanguageCa
ca

AnalysisSchemeLanguageCs
cs

AnalysisSchemeLanguageDa
da

AnalysisSchemeLanguageDe
de

AnalysisSchemeLanguageEl
el

AnalysisSchemeLanguageEn
en

AnalysisSchemeLanguageEs
es

AnalysisSchemeLanguageEu
eu

AnalysisSchemeLanguageFa
fa

AnalysisSchemeLanguageFi
fi

AnalysisSchemeLanguageFr
fr

AnalysisSchemeLanguageGa
ga

AnalysisSchemeLanguageGl
gl

AnalysisSchemeLanguageHe
he

AnalysisSchemeLanguageHi
hi

AnalysisSchemeLanguageHu
hu

AnalysisSchemeLanguageHy
hy

AnalysisSchemeLanguageId
id

AnalysisSchemeLanguageIt
it

AnalysisSchemeLanguageJa
ja

AnalysisSchemeLanguageKo
ko

AnalysisSchemeLanguageLv
lv

AnalysisSchemeLanguageMul
mul

AnalysisSchemeLanguageNl
nl

AnalysisSchemeLanguageNo
no

AnalysisSchemeLanguagePt
pt

AnalysisSchemeLanguageRo
ro

AnalysisSchemeLanguageRu
ru

AnalysisSchemeLanguageSv
sv

AnalysisSchemeLanguageTh
th

AnalysisSchemeLanguageTr
tr

AnalysisSchemeLanguageZhHans
zh-Hans

AnalysisSchemeLanguageZhHant
zh-Hant

IndexFieldType

data IndexFieldType

Constructors

IndexFieldTypeDate
date

IndexFieldTypeDateArray
date-array

IndexFieldTypeDouble
double

IndexFieldTypeDoubleArray
double-array

IndexFieldTypeInt
int

IndexFieldTypeIntArray
int-array

IndexFieldTypeLatlon
latlon

IndexFieldTypeLiteral
literal

IndexFieldTypeLiteralArray
literal-array

IndexFieldTypeText
text

IndexFieldTypeTextArray
text-array

OptionState

data OptionState

Constructors

OptionStateActive
Active

OptionStateFailedToValidate
FailedToValidate

OptionStateProcessing
Processing

OptionStateRequiresIndexDocuments
RequiresIndexDocuments

PartitionInstanceType

data PartitionInstanceType

Constructors

PartitionInstanceTypeSearchM1Large
search.m1.large

PartitionInstanceTypeSearchM1Small
search.m1.small

PartitionInstanceTypeSearchM22Xlarge
search.m2.2xlarge

PartitionInstanceTypeSearchM2Xlarge
search.m2.xlarge

SuggesterFuzzyMatching

data SuggesterFuzzyMatching

Constructors

SuggesterFuzzyMatchingHigh
high

SuggesterFuzzyMatchingLow
low

SuggesterFuzzyMatchingNone
none

ServiceEndpoint

data ServiceEndpoint

The service endpoint for updating documents in a search domain.

serviceEndpoint :: ServiceEndpoint

Smart constructor for the minimum required fields to construct a valid ServiceEndpoint data type to populate a request.

The fields accessible through corresponding lenses are:

Endpoint :: Maybe Text

seEndpoint :: Lens' ServiceEndpoint (Maybe Text)

The endpoint to which service requests can be submitted. For example, search-imdb-movies-oopcnjfn6ugofer3zx5iadxxca.eu-west-1.cloudsearch.amazonaws.com or doc-imdb-movies-oopcnjfn6ugofer3zx5iadxxca.eu-west-1.cloudsearch.amazonaws.com.

AccessPoliciesStatus

data AccessPoliciesStatus

The access rules configured for the domain specified in the request.

accessPoliciesStatus

Arguments

:: Text
apsOptions

-> OptionStatus
apsStatus

-> AccessPoliciesStatus

Smart constructor for the minimum required fields to construct a valid AccessPoliciesStatus data type.

AccessPoliciesStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: Text

Status :: OptionStatus

apsOptions :: Lens' AccessPoliciesStatus Text

Access rules for a domain's document or search service endpoints. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide. The maximum size of a policy document is 100 KB.

apsStatus :: Lens' AccessPoliciesStatus OptionStatus

The status of domain configuration option.

AnalysisOptions

data AnalysisOptions

Synonyms, stopwords, and stemming options for an analysis scheme.

analysisOptions :: AnalysisOptions

Smart constructor for the minimum required fields to construct a valid AnalysisOptions data type to populate a request.

The fields accessible through corresponding lenses are:

Synonyms :: Maybe Text

Stopwords :: Maybe Text

StemmingDictionary :: Maybe Text

AlgorithmicStemming :: Maybe AlgorithmicStemming

aoSynonyms :: Lens' AnalysisOptions (Maybe Text)

A JSON object that defines synonym groups and aliases. A synonym group is an array of arrays, where each sub-array is a group of terms where each term in the group is considered a synonym of every other term in the group. The aliases value is an object that contains a collection of string:value pairs where the string specifies a term and the array of values specifies each of the aliases for that term. An alias is considered a synonym of the specified term, but the term is not considered a synonym of the alias. For more information about specifying synonyms, see Synonyms in the Amazon CloudSearch Developer Guide.

aoStopwords :: Lens' AnalysisOptions (Maybe Text)

A JSON array of terms to ignore during indexing and searching. For example, ["a", "an", "the", "of"]. The stopwords dictionary must explicitly list each word you want to ignore. Wildcards and regular expressions are not supported.

aoStemmingDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON object that contains a collection of string:value pairs that each map a term to its stem. For example, {"term1": "stem1", "term2": "stem2", "term3": "stem3"}. The stemming dictionary is applied in addition to any algorithmic stemming. This enables you to override the results of the algorithmic stemming to correct specific cases of overstemming or understemming. The maximum size of a stemming dictionary is 500 KB.

aoAlgorithmicStemming :: Lens' AnalysisOptions (Maybe AlgorithmicStemming)

The level of algorithmic stemming to perform: none, minimal, light, or full. The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearch Developer Guide.

AnalysisScheme

data AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

analysisScheme

Arguments

:: Text
asAnalysisSchemeName

-> AnalysisSchemeLanguage
asAnalysisSchemeLanguage

-> AnalysisScheme

Smart constructor for the minimum required fields to construct a valid AnalysisScheme data type to populate a request.

The fields accessible through corresponding lenses are:

AnalysisSchemeName :: Text

AnalysisSchemeLanguage :: AnalysisSchemeLanguage

AnalysisOptions :: Maybe AnalysisOptions

asAnalysisSchemeName :: Lens' AnalysisScheme Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

asAnalysisSchemeLanguage :: Lens' AnalysisScheme AnalysisSchemeLanguage

An IETF RFC 4646 language code or mul for multiple languages.

asAnalysisOptions :: Lens' AnalysisScheme (Maybe AnalysisOptions)

Synonyms, stopwords, and stemming options for an analysis scheme.

AnalysisSchemeStatus

data AnalysisSchemeStatus

The status and configuration of an AnalysisScheme.

analysisSchemeStatus

Arguments

:: AnalysisScheme
assOptions

-> OptionStatus
assStatus

-> AnalysisSchemeStatus

Smart constructor for the minimum required fields to construct a valid AnalysisSchemeStatus data type.

AnalysisSchemeStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: AnalysisScheme

Status :: OptionStatus

assOptions :: Lens' AnalysisSchemeStatus AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

assStatus :: Lens' AnalysisSchemeStatus OptionStatus

The status of domain configuration option.

AvailabilityOptionsStatus

data AvailabilityOptionsStatus

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

availabilityOptionsStatus

Arguments

:: Bool
aosOptions

-> OptionStatus
aosStatus

-> AvailabilityOptionsStatus

Smart constructor for the minimum required fields to construct a valid AvailabilityOptionsStatus data type.

AvailabilityOptionsStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: Bool

Status :: OptionStatus

aosOptions :: Lens' AvailabilityOptionsStatus Bool

The availability options configured for the domain.

aosStatus :: Lens' AvailabilityOptionsStatus OptionStatus

The status of domain configuration option.

DateArrayOptions

data DateArrayOptions

Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array. All options are enabled by default.

dateArrayOptions :: DateArrayOptions

Smart constructor for the minimum required fields to construct a valid DateArrayOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceFields :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

dao1DefaultValue :: Lens' DateArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

dao1SourceFields :: Lens' DateArrayOptions (Maybe Text)

A list of source fields to map to the field.

dao1FacetEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

dao1SearchEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

dao1ReturnEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

DateOptions

data DateOptions

Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date. All options are enabled by default.

dateOptions :: DateOptions

Smart constructor for the minimum required fields to construct a valid DateOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceField :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

SortEnabled :: Maybe Bool

do1DefaultValue :: Lens' DateOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

do1SourceField :: Lens' DateOptions (Maybe Text)

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

do1FacetEnabled :: Lens' DateOptions (Maybe Bool)

Whether facet information can be returned for the field.

do1SearchEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field are searchable.

do1ReturnEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

do1SortEnabled :: Lens' DateOptions (Maybe Bool)

Whether the field can be used to sort the search results.

DocumentSuggesterOptions

data DocumentSuggesterOptions

Options for a search suggester.

documentSuggesterOptions

Arguments

:: Text
dsoSourceField

-> DocumentSuggesterOptions

Smart constructor for the minimum required fields to construct a valid DocumentSuggesterOptions data type to populate a request.

The fields accessible through corresponding lenses are:

SourceField :: Text

FuzzyMatching :: Maybe SuggesterFuzzyMatching

SortExpression :: Maybe Text

dsoSourceField :: Lens' DocumentSuggesterOptions Text

The name of the index field you want to use for suggestions.

dsoFuzzyMatching :: Lens' DocumentSuggesterOptions (Maybe SuggesterFuzzyMatching)

The level of fuzziness allowed when suggesting matches for a string: none, low, or high. With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

dsoSortExpression :: Lens' DocumentSuggesterOptions (Maybe Text)

An expression that computes a score for each suggestion to control how they are sorted.

DomainStatus

data DomainStatus

The current status of the search domain.

domainStatus

Arguments

:: Text
dsDomainId

-> Text
dsDomainName

-> Bool
dsRequiresIndexDocuments

-> DomainStatus

Smart constructor for the minimum required fields to construct a valid DomainStatus data type.

DomainStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DomainId :: Text

DomainName :: Text

ARN :: Maybe Text

Created :: Maybe Bool

Deleted :: Maybe Bool

DocService :: Maybe ServiceEndpoint

SearchService :: Maybe ServiceEndpoint

RequiresIndexDocuments :: Bool

Processing :: Maybe Bool

SearchInstanceType :: Maybe Text

SearchPartitionCount :: Maybe Integer

SearchInstanceCount :: Maybe Integer

dsDomainId :: Lens' DomainStatus Text

An internally generated unique identifier for a domain.

dsDomainName :: Lens' DomainStatus Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dsARN :: Lens' DomainStatus (Maybe Text)

The Amazon Resource Name (ARN) of the search domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

dsCreated :: Lens' DomainStatus (Maybe Bool)

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

dsDeleted :: Lens' DomainStatus (Maybe Bool)

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

dsDocService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for updating documents in a search domain.

dsSearchService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for requesting search results from a search domain.

dsRequiresIndexDocuments :: Lens' DomainStatus Bool

True if IndexDocuments needs to be called to activate the current domain configuration.

dsProcessing :: Lens' DomainStatus (Maybe Bool)

True if processing is being done to activate the current domain configuration.

dsSearchInstanceType :: Lens' DomainStatus (Maybe Text)

The instance type that is being used to process search requests.

dsSearchPartitionCount :: Lens' DomainStatus (Maybe Integer)

The number of partitions across which the search index is spread.

dsSearchInstanceCount :: Lens' DomainStatus (Maybe Integer)

The number of search instances that are available to process search requests.

DoubleArrayOptions

data DoubleArrayOptions

Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array. All options are enabled by default.

doubleArrayOptions :: DoubleArrayOptions

Smart constructor for the minimum required fields to construct a valid DoubleArrayOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Double

SourceFields :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

daoDefaultValue :: Lens' DoubleArrayOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document.

daoSourceFields :: Lens' DoubleArrayOptions (Maybe Text)

A list of source fields to map to the field.

daoFacetEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

daoSearchEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

daoReturnEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

DoubleOptions

data DoubleOptions

Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double. All options are enabled by default.

doubleOptions :: DoubleOptions

Smart constructor for the minimum required fields to construct a valid DoubleOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Double

SourceField :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

SortEnabled :: Maybe Bool

doDefaultValue :: Lens' DoubleOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

doSourceField :: Lens' DoubleOptions (Maybe Text)

The name of the source field to map to the field.

doFacetEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether facet information can be returned for the field.

doSearchEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field are searchable.

doReturnEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

doSortEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the field can be used to sort the search results.

Expression

data Expression

A named expression that can be evaluated at search time. Can be used for sorting and filtering search results and constructing other expressions.

expression

Arguments

:: Text
eExpressionName

-> Text
eExpressionValue

-> Expression

Smart constructor for the minimum required fields to construct a valid Expression data type to populate a request.

The fields accessible through corresponding lenses are:

ExpressionName :: Text

ExpressionValue :: Text

eExpressionName :: Lens' Expression Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

eExpressionValue :: Lens' Expression Text

The expression to evaluate for sorting while processing a search request. The Expression syntax is based on JavaScript expressions. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

ExpressionStatus

data ExpressionStatus

The value of an Expression and its current status.

expressionStatus

Arguments

:: Expression
esOptions

-> OptionStatus
esStatus

-> ExpressionStatus

Smart constructor for the minimum required fields to construct a valid ExpressionStatus data type.

ExpressionStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: Expression

Status :: OptionStatus

esOptions :: Lens' ExpressionStatus Expression

The expression that is evaluated for sorting or filtering while processing a search request.

esStatus :: Lens' ExpressionStatus OptionStatus

The status of domain configuration option.

IndexField

data IndexField

The index field and field options you want to configure.

indexField

Arguments

:: Text
ifIndexFieldName

-> IndexFieldType
ifIndexFieldType

-> IndexField

Smart constructor for the minimum required fields to construct a valid IndexField data type to populate a request.

The fields accessible through corresponding lenses are:

IndexFieldName :: Text

IndexFieldType :: IndexFieldType

IntOptions :: Maybe IntOptions

DoubleOptions :: Maybe DoubleOptions

LiteralOptions :: Maybe LiteralOptions

TextOptions :: Maybe TextOptions

DateOptions :: Maybe DateOptions

LatLonOptions :: Maybe LatLonOptions

IntArrayOptions :: Maybe IntArrayOptions

DoubleArrayOptions :: Maybe DoubleArrayOptions

LiteralArrayOptions :: Maybe LiteralArrayOptions

TextArrayOptions :: Maybe TextArrayOptions

DateArrayOptions :: Maybe DateArrayOptions

ifIndexFieldName :: Lens' IndexField Text

The name of a field in the search index. Field names must begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Uppercase letters and hyphens are not allowed. The name "score" is reserved and cannot be specified as field or expression name.

ifIndexFieldType :: Lens' IndexField IndexFieldType

The type of field. The valid options for a field depend on the field type. For more information about the supported field types, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

ifIntOptions :: Lens' IndexField (Maybe IntOptions)

Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int. All options are enabled by default.

ifDoubleOptions :: Lens' IndexField (Maybe DoubleOptions)

Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double. All options are enabled by default.

ifLiteralOptions :: Lens' IndexField (Maybe LiteralOptions)

Options for literal field. Present if IndexFieldType specifies the field is of type literal. All options are enabled by default.

ifTextOptions :: Lens' IndexField (Maybe TextOptions)

Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

ifDateOptions :: Lens' IndexField (Maybe DateOptions)

Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date. All options are enabled by default.

ifLatLonOptions :: Lens' IndexField (Maybe LatLonOptions)

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

ifIntArrayOptions :: Lens' IndexField (Maybe IntArrayOptions)

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

ifDoubleArrayOptions :: Lens' IndexField (Maybe DoubleArrayOptions)

Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array. All options are enabled by default.

ifLiteralArrayOptions :: Lens' IndexField (Maybe LiteralArrayOptions)

Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array. All options are enabled by default.

ifTextArrayOptions :: Lens' IndexField (Maybe TextArrayOptions)

Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array. A text-array field is always searchable. All options are enabled by default.

ifDateArrayOptions :: Lens' IndexField (Maybe DateArrayOptions)

Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array. All options are enabled by default.

IndexFieldStatus

data IndexFieldStatus

The value of an IndexField and its current status.

indexFieldStatus

Arguments

:: IndexField
ifsOptions

-> OptionStatus
ifsStatus

-> IndexFieldStatus

Smart constructor for the minimum required fields to construct a valid IndexFieldStatus data type.

IndexFieldStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: IndexField

Status :: OptionStatus

ifsOptions :: Lens' IndexFieldStatus IndexField

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

ifsStatus :: Lens' IndexFieldStatus OptionStatus

The status of domain configuration option.

IntArrayOptions

data IntArrayOptions

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

intArrayOptions :: IntArrayOptions

Smart constructor for the minimum required fields to construct a valid IntArrayOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Integer

SourceFields :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

iaoDefaultValue :: Lens' IntArrayOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document.

iaoSourceFields :: Lens' IntArrayOptions (Maybe Text)

A list of source fields to map to the field.

iaoFacetEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

iaoSearchEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

iaoReturnEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

IntOptions

data IntOptions

Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int. All options are enabled by default.

intOptions :: IntOptions

Smart constructor for the minimum required fields to construct a valid IntOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Integer

SourceField :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

SortEnabled :: Maybe Bool

ioDefaultValue :: Lens' IntOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

ioSourceField :: Lens' IntOptions (Maybe Text)

The name of the source field to map to the field.

ioFacetEnabled :: Lens' IntOptions (Maybe Bool)

Whether facet information can be returned for the field.

ioSearchEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field are searchable.

ioReturnEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

ioSortEnabled :: Lens' IntOptions (Maybe Bool)

Whether the field can be used to sort the search results.

LatLonOptions

data LatLonOptions

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

latLonOptions :: LatLonOptions

Smart constructor for the minimum required fields to construct a valid LatLonOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceField :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

SortEnabled :: Maybe Bool

lloDefaultValue :: Lens' LatLonOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

lloSourceField :: Lens' LatLonOptions (Maybe Text)

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

lloFacetEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether facet information can be returned for the field.

lloSearchEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field are searchable.

lloReturnEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

lloSortEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the field can be used to sort the search results.

LiteralArrayOptions

data LiteralArrayOptions

Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array. All options are enabled by default.

literalArrayOptions :: LiteralArrayOptions

Smart constructor for the minimum required fields to construct a valid LiteralArrayOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceFields :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

laoDefaultValue :: Lens' LiteralArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

laoSourceFields :: Lens' LiteralArrayOptions (Maybe Text)

A list of source fields to map to the field.

laoFacetEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

laoSearchEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

laoReturnEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

LiteralOptions

data LiteralOptions

Options for literal field. Present if IndexFieldType specifies the field is of type literal. All options are enabled by default.

literalOptions :: LiteralOptions

Smart constructor for the minimum required fields to construct a valid LiteralOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceField :: Maybe Text

FacetEnabled :: Maybe Bool

SearchEnabled :: Maybe Bool

ReturnEnabled :: Maybe Bool

SortEnabled :: Maybe Bool

loDefaultValue :: Lens' LiteralOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

loSourceField :: Lens' LiteralOptions (Maybe Text)

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

loFacetEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether facet information can be returned for the field.

loSearchEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field are searchable.

loReturnEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

loSortEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the field can be used to sort the search results.

OptionStatus

data OptionStatus

The status of domain configuration option.

optionStatus

Arguments

:: ISO8601
osCreationDate

-> ISO8601
osUpdateDate

-> OptionState
osState

-> OptionStatus

Smart constructor for the minimum required fields to construct a valid OptionStatus data type to populate a request.

The fields accessible through corresponding lenses are:

CreationDate :: ISO8601

UpdateDate :: ISO8601

UpdateVersion :: Maybe Integer

State :: OptionState

PendingDeletion :: Maybe Bool

osCreationDate :: Lens' OptionStatus ISO8601

A timestamp for when this option was created.

osUpdateDate :: Lens' OptionStatus ISO8601

A timestamp for when this option was last updated.

osUpdateVersion :: Lens' OptionStatus (Maybe Integer)

A unique integer that indicates when this option was last updated.

osState :: Lens' OptionStatus OptionState

The state of processing a change to an option. Possible values: RequiresIndexDocuments: the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. Processing: the option's latest value is in the process of being activated. Active: the option's latest value is completely deployed. FailedToValidate: the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

osPendingDeletion :: Lens' OptionStatus (Maybe Bool)

Indicates that the option will be deleted once processing is complete.

ScalingParameters

data ScalingParameters

The desired instance type and desired number of replicas of each index partition.

scalingParameters :: ScalingParameters

Smart constructor for the minimum required fields to construct a valid ScalingParameters data type to populate a request.

The fields accessible through corresponding lenses are:

DesiredInstanceType :: Maybe PartitionInstanceType

DesiredReplicationCount :: Maybe Integer

DesiredPartitionCount :: Maybe Integer

spDesiredInstanceType :: Lens' ScalingParameters (Maybe PartitionInstanceType)

The instance type that you want to preconfigure for your domain. For example, search.m1.small.

spDesiredReplicationCount :: Lens' ScalingParameters (Maybe Integer)

The number of replicas you want to preconfigure for each index partition.

spDesiredPartitionCount :: Lens' ScalingParameters (Maybe Integer)

The number of partitions you want to preconfigure for your domain. Only valid when you select m2.2xlarge as the desired instance type.

ScalingParametersStatus

data ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

scalingParametersStatus

Arguments

:: ScalingParameters
spsOptions

-> OptionStatus
spsStatus

-> ScalingParametersStatus

Smart constructor for the minimum required fields to construct a valid ScalingParametersStatus data type.

ScalingParametersStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: ScalingParameters

Status :: OptionStatus

spsOptions :: Lens' ScalingParametersStatus ScalingParameters

The desired instance type and desired number of replicas of each index partition.

spsStatus :: Lens' ScalingParametersStatus OptionStatus

The status of domain configuration option.

Suggester

data Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

suggester

Arguments

:: Text
sSuggesterName

-> DocumentSuggesterOptions
sDocumentSuggesterOptions

-> Suggester

Smart constructor for the minimum required fields to construct a valid Suggester data type to populate a request.

The fields accessible through corresponding lenses are:

SuggesterName :: Text

DocumentSuggesterOptions :: DocumentSuggesterOptions

sSuggesterName :: Lens' Suggester Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

sDocumentSuggesterOptions :: Lens' Suggester DocumentSuggesterOptions

Options for a search suggester.

SuggesterStatus

data SuggesterStatus

The value of a Suggester and its current status.

suggesterStatus

Arguments

:: Suggester
ssOptions

-> OptionStatus
ssStatus

-> SuggesterStatus

Smart constructor for the minimum required fields to construct a valid SuggesterStatus data type.

SuggesterStatus is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Options :: Suggester

Status :: OptionStatus

ssOptions :: Lens' SuggesterStatus Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

ssStatus :: Lens' SuggesterStatus OptionStatus

The status of domain configuration option.

TextArrayOptions

data TextArrayOptions

Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array. A text-array field is always searchable. All options are enabled by default.

textArrayOptions :: TextArrayOptions

Smart constructor for the minimum required fields to construct a valid TextArrayOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceFields :: Maybe Text

ReturnEnabled :: Maybe Bool

HighlightEnabled :: Maybe Bool

AnalysisScheme :: Maybe Text

taoDefaultValue :: Lens' TextArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

taoSourceFields :: Lens' TextArrayOptions (Maybe Text)

A list of source fields to map to the field.

taoReturnEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

taoHighlightEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether highlights can be returned for the field.

taoAnalysisScheme :: Lens' TextArrayOptions (Maybe Text)

The name of an analysis scheme for a text-array field.

TextOptions

data TextOptions

Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

textOptions :: TextOptions

Smart constructor for the minimum required fields to construct a valid TextOptions data type to populate a request.

The fields accessible through corresponding lenses are:

DefaultValue :: Maybe Text

SourceField :: Maybe Text

ReturnEnabled :: Maybe Bool

SortEnabled :: Maybe Bool

HighlightEnabled :: Maybe Bool

AnalysisScheme :: Maybe Text

toDefaultValue :: Lens' TextOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

toSourceField :: Lens' TextOptions (Maybe Text)

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

toReturnEnabled :: Lens' TextOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

toSortEnabled :: Lens' TextOptions (Maybe Bool)

Whether the field can be used to sort the search results.

toHighlightEnabled :: Lens' TextOptions (Maybe Bool)

Whether highlights can be returned for the field.

toAnalysisScheme :: Lens' TextOptions (Maybe Text)

The name of an analysis scheme for a text field.