Network.AWS.CloudSearch.UpdateAvailabilityOptions

Description

Configures the availability options for a domain. Enabling the Multi-AZ option expands an Amazon CloudSearch domain to an additional Availability Zone in the same Region to increase fault tolerance in the event of a service disruption. Changes to the Multi-AZ option can take about half an hour to become active. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_UpdateAvailabilityOptions.html

Request

data UpdateAvailabilityOptions

Request constructor

updateAvailabilityOptions

Arguments

:: Text
uaoDomainName

-> Bool
uaoMultiAZ

-> UpdateAvailabilityOptions

UpdateAvailabilityOptions constructor.

The fields accessible through corresponding lenses are:

uaoDomainName :: Text

uaoMultiAZ :: Bool

Request lenses

uaoDomainName :: Lens' UpdateAvailabilityOptions Text

uaoMultiAZ :: Lens' UpdateAvailabilityOptions Bool

You expand an existing search domain to a second Availability Zone by setting the Multi-AZ option to true. Similarly, you can turn off the Multi-AZ option to downgrade the domain to a single Availability Zone by setting the Multi-AZ option to false.

Response

data UpdateAvailabilityOptionsResponse

Response constructor

updateAvailabilityOptionsResponse :: UpdateAvailabilityOptionsResponse

UpdateAvailabilityOptionsResponse constructor.

The fields accessible through corresponding lenses are:

uaorAvailabilityOptions :: Maybe AvailabilityOptionsStatus

Response lenses

uaorAvailabilityOptions :: Lens' UpdateAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

The newly-configured availability options. Indicates whether Multi-AZ is enabled for the domain.