Network.AWS.CloudSearch.UpdateScalingParameters

Description

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_UpdateScalingParameters.html

Request

data UpdateScalingParameters

Request constructor

updateScalingParameters

Arguments

:: Text
uspDomainName

-> ScalingParameters
uspScalingParameters

-> UpdateScalingParameters

UpdateScalingParameters constructor.

The fields accessible through corresponding lenses are:

uspDomainName :: Text

uspScalingParameters :: ScalingParameters

Request lenses

uspDomainName :: Lens' UpdateScalingParameters Text

uspScalingParameters :: Lens' UpdateScalingParameters ScalingParameters

Response

data UpdateScalingParametersResponse

Response constructor

updateScalingParametersResponse

Arguments

:: ScalingParametersStatus
usprScalingParameters

-> UpdateScalingParametersResponse

UpdateScalingParametersResponse constructor.

The fields accessible through corresponding lenses are:

usprScalingParameters :: ScalingParametersStatus

Response lenses

usprScalingParameters :: Lens' UpdateScalingParametersResponse ScalingParametersStatus