Network.AWS.CloudSearch.UpdateServiceAccessPolicies

Description

Configures the access rules that control access to the domain's document and search endpoints. For more information, see Configuring Access for an Amazon CloudSearch Domain.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_UpdateServiceAccessPolicies.html

Request

data UpdateServiceAccessPolicies

Request constructor

updateServiceAccessPolicies

Arguments

:: Text
usapDomainName

-> Text
usapAccessPolicies

-> UpdateServiceAccessPolicies

UpdateServiceAccessPolicies constructor.

The fields accessible through corresponding lenses are:

usapAccessPolicies :: Text

usapDomainName :: Text

Request lenses

usapAccessPolicies :: Lens' UpdateServiceAccessPolicies Text

The access rules you want to configure. These rules replace any existing rules.

usapDomainName :: Lens' UpdateServiceAccessPolicies Text

Response

data UpdateServiceAccessPoliciesResponse

Response constructor

updateServiceAccessPoliciesResponse

Arguments

:: AccessPoliciesStatus
usaprAccessPolicies

-> UpdateServiceAccessPoliciesResponse

UpdateServiceAccessPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

usaprAccessPolicies :: AccessPoliciesStatus

Response lenses

usaprAccessPolicies :: Lens' UpdateServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain.