Network.AWS.CloudSearch.V2013_01_01

Description

Amazon CloudSearch is a fully-managed service in the AWS Cloud that makes it simple and cost-effective to set up, manage, and scale a search solution for your website or application. Amazon CloudSearch supports 34 languages and popular search features such as highlighting, autocomplete, and geospatial search.

Documentation

module Network.AWS.CloudSearch.V2013_01_01.BuildSuggesters

module Network.AWS.CloudSearch.V2013_01_01.CreateDomain

module Network.AWS.CloudSearch.V2013_01_01.DefineAnalysisScheme

module Network.AWS.CloudSearch.V2013_01_01.DefineExpression

module Network.AWS.CloudSearch.V2013_01_01.DefineIndexField

module Network.AWS.CloudSearch.V2013_01_01.DefineSuggester

module Network.AWS.CloudSearch.V2013_01_01.DeleteAnalysisScheme

module Network.AWS.CloudSearch.V2013_01_01.DeleteDomain

module Network.AWS.CloudSearch.V2013_01_01.DeleteExpression

module Network.AWS.CloudSearch.V2013_01_01.DeleteIndexField

module Network.AWS.CloudSearch.V2013_01_01.DeleteSuggester

module Network.AWS.CloudSearch.V2013_01_01.DescribeAnalysisSchemes

module Network.AWS.CloudSearch.V2013_01_01.DescribeAvailabilityOptions

module Network.AWS.CloudSearch.V2013_01_01.DescribeDomains

module Network.AWS.CloudSearch.V2013_01_01.DescribeExpressions

module Network.AWS.CloudSearch.V2013_01_01.DescribeIndexFields

module Network.AWS.CloudSearch.V2013_01_01.DescribeScalingParameters

module Network.AWS.CloudSearch.V2013_01_01.DescribeServiceAccessPolicies

module Network.AWS.CloudSearch.V2013_01_01.DescribeSuggesters

module Network.AWS.CloudSearch.V2013_01_01.IndexDocuments

module Network.AWS.CloudSearch.V2013_01_01.ListDomainNames

module Network.AWS.CloudSearch.V2013_01_01.Types

module Network.AWS.CloudSearch.V2013_01_01.UpdateAvailabilityOptions

module Network.AWS.CloudSearch.V2013_01_01.UpdateScalingParameters

module Network.AWS.CloudSearch.V2013_01_01.UpdateServiceAccessPolicies