Network.AWS.CloudSearch.V2013_01_01.DefineExpression

Description

Configures an Expression for the search domain. Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Documentation

data DefineExpression

Constructors

DefineExpression

Fields

_detDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_detExpression :: Expression

A named expression that can be evaluated at search time. Can be used for sorting and filtering search results and constructing other expressions.

detExpression :: Lens' DefineExpression Expression

detDomainName :: Lens' DefineExpression Text

data DefineExpressionResponse

Constructors

DefineExpressionResponse

Fields

_deuExpression :: ExpressionStatus

The value of an Expression and its current status.

deuExpression :: Iso' DefineExpressionResponse ExpressionStatus