Network.AWS.CloudSearch.V2013_01_01.DeleteAnalysisScheme

Description

Deletes an analysis scheme. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Request

data DeleteAnalysisScheme

Container for the parameters to the DeleteAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme you want to delete.

Request constructor

mkDeleteAnalysisScheme

Arguments

:: Text
das1DomainName

-> Text
das1AnalysisSchemeName

-> DeleteAnalysisScheme

Smart constructor for the minimum required parameters to construct a valid DeleteAnalysisScheme request.

The fields accessible through corresponding lenses are:

DomainName :: Text

AnalysisSchemeName :: Text

Request lenses

das1DomainName :: Lens' DeleteAnalysisScheme Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

das1AnalysisSchemeName :: Lens' DeleteAnalysisScheme Text

The name of the analysis scheme you want to delete.

Response

data DeleteAnalysisSchemeResponse

The result of a DeleteAnalysisScheme request. Contains the status of the deleted analysis scheme.

Response constructor

mkDeleteAnalysisSchemeResponse

Arguments

:: AnalysisSchemeStatus
dasrrAnalysisScheme

-> DeleteAnalysisSchemeResponse

Smart constructor for the minimum required parameters to construct a valid DeleteAnalysisSchemeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AnalysisScheme :: AnalysisSchemeStatus

Response lenses

dasrrAnalysisScheme :: Lens' DeleteAnalysisSchemeResponse AnalysisSchemeStatus

The status of the analysis scheme being deleted.