Network.AWS.CloudSearch.V2013_01_01.DeleteDomain

Description

Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see Deleting a Search Domain in the Amazon CloudSearch Developer Guide.

Documentation

data DeleteDomain

Constructors

DeleteDomain

Fields

_ddyDomainName :: Text

The name of the domain you want to permanently delete.

ddyDomainName :: Iso' DeleteDomain Text

data DeleteDomainResponse

Constructors

DeleteDomainResponse

Fields

_ddzDomainStatus :: Maybe DomainStatus

The current status of the search domain.

ddzDomainStatus :: Iso' DeleteDomainResponse (Maybe DomainStatus)