Network.AWS.CloudSearch.V2013_01_01.DeleteIndexField

Description

Removes an IndexField from the search domain. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Documentation

data DeleteIndexField

Constructors

DeleteIndexField

Fields

_diftDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_diftIndexFieldName :: Text

The name of the index field your want to remove from the domain's indexing options.

diftIndexFieldName :: Lens' DeleteIndexField Text

diftDomainName :: Lens' DeleteIndexField Text

data DeleteIndexFieldResponse

Constructors

DeleteIndexFieldResponse

Fields

_difuIndexField :: IndexFieldStatus

The status of the index field being deleted.

difuIndexField :: Iso' DeleteIndexFieldResponse IndexFieldStatus