Network.AWS.CloudSearch.V2013_01_01.DeleteSuggester

Description

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Request

data DeleteSuggester

Container for the parameters to the DeleteSuggester operation. Specifies the name of the domain you want to update and name of the suggester you want to delete.

Request constructor

mkDeleteSuggester

Arguments

:: Text
ds2DomainName

-> Text
ds2SuggesterName

-> DeleteSuggester

Smart constructor for the minimum required parameters to construct a valid DeleteSuggester request.

Request lenses

ds2DomainName :: Lens' DeleteSuggester Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

ds2SuggesterName :: Lens' DeleteSuggester Text

Specifies the name of the suggester you want to delete.

Response

data DeleteSuggesterResponse

The result of a DeleteSuggester request. Contains the status of the deleted suggester.

Response lenses

dsrrSuggester :: Lens' DeleteSuggesterResponse SuggesterStatus

The status of the suggester being deleted.