Network.AWS.CloudSearch.V2013_01_01.DescribeDomains

Description

Gets information about the search domains owned by this account. Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: q=matchall&amp;q.parser=structured&amp;size=0. For more information, see Getting Information about a Search Domain in the Amazon CloudSearch Developer Guide.

Request

data DescribeDomains

Container for the parameters to the DescribeDomains operation. By default shows the status of all domains. To restrict the response to particular domains, specify the names of the domains you want to describe.

Request constructor

mkDescribeDomains :: DescribeDomains

Smart constructor for the minimum required parameters to construct a valid DescribeDomains request.

Request lenses

dd1DomainNames :: Lens' DescribeDomains [Text]

The names of the domains you want to include in the response.

Response

data DescribeDomainsResponse

The result of a DescribeDomains request. Contains the status of the domains specified in the request or all domains owned by the account.

Response lenses

ddrsrsDomainStatusList :: Lens' DescribeDomainsResponse [DomainStatus]

A list that contains the status of each requested domain.