Network.AWS.CloudSearch.V2013_01_01.DescribeExpressions

Description

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Request

data DescribeExpressions

Container for the parameters to the DescribeDomains operation. Specifies the name of the domain you want to describe. To restrict the response to particular expressions, specify the names of the expressions you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

Request constructor

mkDescribeExpressions

Arguments

:: Text
de2DomainName

-> DescribeExpressions

Smart constructor for the minimum required parameters to construct a valid DescribeExpressions request.

The fields accessible through corresponding lenses are:

DomainName :: Text

ExpressionNames :: [Text]

Deployed :: Maybe Bool

Request lenses

de2DomainName :: Lens' DescribeExpressions Text

The name of the domain you want to describe.

de2ExpressionNames :: Lens' DescribeExpressions [Text]

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

de2Deployed :: Lens' DescribeExpressions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Response

data DescribeExpressionsResponse

The result of a DescribeExpressions request. Contains the expressions configured for the domain specified in the request.

Response constructor

mkDescribeExpressionsResponse

Arguments

:: [ExpressionStatus]
der1Expressions

-> DescribeExpressionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeExpressionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Expressions :: [ExpressionStatus]

Response lenses

der1Expressions :: Lens' DescribeExpressionsResponse [ExpressionStatus]

The expressions configured for the domain.