Network.AWS.CloudSearch.V2013_01_01.DescribeIndexFields

Description

Gets information about the index fields configured for the search domain. Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Domain Information in the Amazon CloudSearch Developer Guide.

Request

data DescribeIndexFields

Container for the parameters to the DescribeIndexFields operation. Specifies the name of the domain you want to describe. To restrict the response to particular index fields, specify the names of the index fields you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

Request constructor

mkDescribeIndexFields

Arguments

:: Text
dif2DomainName

-> DescribeIndexFields

Smart constructor for the minimum required parameters to construct a valid DescribeIndexFields request.

The fields accessible through corresponding lenses are:

DomainName :: Text

FieldNames :: [Text]

Deployed :: Maybe Bool

Request lenses

dif2DomainName :: Lens' DescribeIndexFields Text

The name of the domain you want to describe.

dif2FieldNames :: Lens' DescribeIndexFields [Text]

A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

dif2Deployed :: Lens' DescribeIndexFields (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Response

data DescribeIndexFieldsResponse

The result of a DescribeIndexFields request. Contains the index fields configured for the domain specified in the request.

Response constructor

mkDescribeIndexFieldsResponse

Arguments

:: [IndexFieldStatus]
difr1IndexFields

-> DescribeIndexFieldsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeIndexFieldsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IndexFields :: [IndexFieldStatus]

Response lenses

difr1IndexFields :: Lens' DescribeIndexFieldsResponse [IndexFieldStatus]

The index fields configured for the domain.