Network.AWS.CloudSearch.V2013_01_01.DescribeServiceAccessPolicies

Description

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

Documentation

describeServiceAccessPolicies

Arguments

:: Text
_dsaprDomainName

-> DescribeServiceAccessPolicies

Minimum specification for a DescribeServiceAccessPolicies request.

data DescribeServiceAccessPolicies

Constructors

DescribeServiceAccessPolicies

Fields

_dsaprDomainName :: Text

The name of the domain you want to describe.

_dsaprDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

dsaprDomainName :: Lens' DescribeServiceAccessPolicies Text

dsaprDeployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

data DescribeServiceAccessPoliciesResponse

Constructors

DescribeServiceAccessPoliciesResponse

Fields

_dsapsAccessPolicies :: AccessPoliciesStatus

The access rules configured for the domain specified in the request.

dsapsAccessPolicies :: Iso' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus