Network.AWS.CloudSearch.V2013_01_01.DescribeSuggesters

Description

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Request

data DescribeSuggesters

Container for the parameters to the DescribeSuggester operation. Specifies the name of the domain you want to describe. To restrict the response to particular suggesters, specify the names of the suggesters you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true.

Request constructor

mkDescribeSuggesters

Arguments

:: Text
ds3DomainName

-> DescribeSuggesters

Smart constructor for the minimum required parameters to construct a valid DescribeSuggesters request.

Request lenses

ds3DomainName :: Lens' DescribeSuggesters Text

The name of the domain you want to describe.

ds3SuggesterNames :: Lens' DescribeSuggesters [Text]

The suggesters you want to describe.

ds3Deployed :: Lens' DescribeSuggesters (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Response

data DescribeSuggestersResponse

The result of a DescribeSuggesters request.

Response lenses

dsrs1Suggesters :: Lens' DescribeSuggestersResponse [SuggesterStatus]

The suggesters configured for the domain specified in the request.