Network.AWS.CloudSearch.V2013_01_01.ListDomainNames

Description

Lists all search domains owned by an account.

Request

data ListDomainNames

Request constructor

mkListDomainNames :: ListDomainNames

Smart constructor for the minimum required parameters to construct a valid ListDomainNames request.

Response

data ListDomainNamesResponse

The result of a ListDomainNames request. Contains a list of the domains owned by an account.

Response constructor

mkListDomainNamesResponse :: ListDomainNamesResponse

Smart constructor for the minimum required parameters to construct a valid ListDomainNamesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

ldnrDomainNames :: Lens' ListDomainNamesResponse (Map Text Text)

The names of the search domains owned by an account.