Network.AWS.CloudSearch.V2013_01_01.UpdateAvailabilityOptions

Description

Configures the availability options for a domain. Enabling the Multi-AZ option expands an Amazon CloudSearch domain to an additional Availability Zone in the same Region to increase fault tolerance in the event of a service disruption. Changes to the Multi-AZ option can take about half an hour to become active. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

Documentation

data UpdateAvailabilityOptions

Constructors

UpdateAvailabilityOptions

Fields

_uaorMultiAZ :: Bool

You expand an existing search domain to a second Availability Zone by setting the Multi-AZ option to true. Similarly, you can turn off the Multi-AZ option to downgrade the domain to a single Availability Zone by setting the Multi-AZ option to false.

_uaorDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

uaorMultiAZ :: Lens' UpdateAvailabilityOptions Bool

uaorDomainName :: Lens' UpdateAvailabilityOptions Text

data UpdateAvailabilityOptionsResponse

Constructors

UpdateAvailabilityOptionsResponse

Fields

_uaosAvailabilityOptions :: Maybe AvailabilityOptionsStatus

The newly-configured availability options. Indicates whether Multi-AZ is enabled for the domain.

uaosAvailabilityOptions :: Iso' UpdateAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)