Network.AWS.CloudSearch.V2013_01_01.UpdateScalingParameters

Description

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

Request

data UpdateScalingParameters

Container for the parameters to the UpdateScalingParameters operation. Specifies the name of the domain you want to update and the scaling parameters you want to configure.

Request constructor

mkUpdateScalingParameters

Arguments

:: Text
uspDomainName

-> ScalingParameters
uspScalingParameters

-> UpdateScalingParameters

Smart constructor for the minimum required parameters to construct a valid UpdateScalingParameters request.

The fields accessible through corresponding lenses are:

DomainName :: Text

ScalingParameters :: ScalingParameters

Request lenses

uspDomainName :: Lens' UpdateScalingParameters Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

uspScalingParameters :: Lens' UpdateScalingParameters ScalingParameters

The desired instance type and desired number of replicas of each index partition.

Response

data UpdateScalingParametersResponse

The result of a UpdateScalingParameters request. Contains the status of the newly-configured scaling parameters.

Response constructor

mkUpdateScalingParametersResponse

Arguments

:: ScalingParametersStatus
usprScalingParameters

-> UpdateScalingParametersResponse

Smart constructor for the minimum required parameters to construct a valid UpdateScalingParametersResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ScalingParameters :: ScalingParametersStatus

Response lenses

usprScalingParameters :: Lens' UpdateScalingParametersResponse ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.