Network.AWS.CloudSearch.CreateDomain

Description

Creates a new search domain. For more information, see Creating a Search Domain in the Amazon CloudSearch Developer Guide .

Creating a Request

createDomain

Arguments

:: Text
cdDomainName

-> CreateDomain

Creates a value of CreateDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdDomainName - A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

data CreateDomain

Container for the parameters to the CreateDomain operation. Specifies a name for the new search domain.

See: createDomain smart constructor.

Request Lenses

cdDomainName :: Lens' CreateDomain Text

A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

Destructuring the Response

createDomainResponse

Arguments

:: Int
cdrsResponseStatus

-> CreateDomainResponse

Creates a value of CreateDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdrsDomainStatus - Undocumented member.

cdrsResponseStatus - -- | The response status code.

data CreateDomainResponse

The result of a CreateDomainRequest . Contains the status of a newly created domain.

See: createDomainResponse smart constructor.

Response Lenses

cdrsDomainStatus :: Lens' CreateDomainResponse (Maybe DomainStatus)

Undocumented member.

cdrsResponseStatus :: Lens' CreateDomainResponse Int

- | The response status code.