Network.AWS.CloudSearch.DefineAnalysisScheme

Description

Configures an analysis scheme that can be applied to a text or 'text-array' field to define language-specific text processing options. For more information, see Configuring Analysis Schemes in the Amazon CloudSearchDeveloper Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DefineAnalysisScheme.html

Request

data DefineAnalysisScheme

Request constructor

defineAnalysisScheme

Arguments

:: Text
das2DomainName

-> AnalysisScheme
das2AnalysisScheme

-> DefineAnalysisScheme

DefineAnalysisScheme constructor.

The fields accessible through corresponding lenses are:

das2AnalysisScheme :: AnalysisScheme

das2DomainName :: Text

Request lenses

das2AnalysisScheme :: Lens' DefineAnalysisScheme AnalysisScheme

das2DomainName :: Lens' DefineAnalysisScheme Text

Response

data DefineAnalysisSchemeResponse

Response constructor

defineAnalysisSchemeResponse

Arguments

:: AnalysisSchemeStatus
dasr1AnalysisScheme

-> DefineAnalysisSchemeResponse

DefineAnalysisSchemeResponse constructor.

The fields accessible through corresponding lenses are:

dasr1AnalysisScheme :: AnalysisSchemeStatus

Response lenses

dasr1AnalysisScheme :: Lens' DefineAnalysisSchemeResponse AnalysisSchemeStatus