Network.AWS.CloudSearch.DefineSuggester

Description

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide .

Creating a Request

defineSuggester

Arguments

:: Text
defDomainName

-> Suggester
defSuggester

-> DefineSuggester

Creates a value of DefineSuggester with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

defDomainName - Undocumented member.

defSuggester - Undocumented member.

data DefineSuggester

Container for the parameters to the DefineSuggester operation. Specifies the name of the domain you want to update and the suggester configuration.

See: defineSuggester smart constructor.

Request Lenses

defDomainName :: Lens' DefineSuggester Text

Undocumented member.

defSuggester :: Lens' DefineSuggester Suggester

Undocumented member.

Destructuring the Response

defineSuggesterResponse

Arguments

:: Int
dsrsResponseStatus

-> SuggesterStatus
dsrsSuggester

-> DefineSuggesterResponse

Creates a value of DefineSuggesterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsResponseStatus - -- | The response status code.

dsrsSuggester - Undocumented member.

data DefineSuggesterResponse

The result of a DefineSuggester request. Contains the status of the newly-configured suggester.

See: defineSuggesterResponse smart constructor.

Response Lenses

dsrsResponseStatus :: Lens' DefineSuggesterResponse Int

- | The response status code.

dsrsSuggester :: Lens' DefineSuggesterResponse SuggesterStatus

Undocumented member.