Network.AWS.CloudSearch.DefineSuggester

Description

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearchDeveloper Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DefineSuggester.html

Request

data DefineSuggester

Request constructor

defineSuggester

Arguments

:: Text
ds2DomainName

-> Suggester
ds2Suggester

-> DefineSuggester

DefineSuggester constructor.

The fields accessible through corresponding lenses are:

ds2DomainName :: Text

ds2Suggester :: Suggester

Request lenses

ds2DomainName :: Lens' DefineSuggester Text

ds2Suggester :: Lens' DefineSuggester Suggester

Response

data DefineSuggesterResponse

Response constructor

defineSuggesterResponse

Arguments

:: SuggesterStatus
dsrSuggester

-> DefineSuggesterResponse

DefineSuggesterResponse constructor.

The fields accessible through corresponding lenses are:

dsrSuggester :: SuggesterStatus

Response lenses

dsrSuggester :: Lens' DefineSuggesterResponse SuggesterStatus