Network.AWS.CloudSearch.DeleteAnalysisScheme

Description

Deletes an analysis scheme. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide .

Creating a Request

deleteAnalysisScheme

Arguments

:: Text
dasDomainName

-> Text
dasAnalysisSchemeName

-> DeleteAnalysisScheme

Creates a value of DeleteAnalysisScheme with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasDomainName - Undocumented member.

dasAnalysisSchemeName - The name of the analysis scheme you want to delete.

data DeleteAnalysisScheme

Container for the parameters to the DeleteAnalysisScheme operation. Specifies the name of the domain you want to update and the analysis scheme you want to delete.

See: deleteAnalysisScheme smart constructor.

Request Lenses

dasDomainName :: Lens' DeleteAnalysisScheme Text

Undocumented member.

dasAnalysisSchemeName :: Lens' DeleteAnalysisScheme Text

The name of the analysis scheme you want to delete.

Destructuring the Response

deleteAnalysisSchemeResponse

Arguments

:: Int
dasarsResponseStatus

-> AnalysisSchemeStatus
dasarsAnalysisScheme

-> DeleteAnalysisSchemeResponse

Creates a value of DeleteAnalysisSchemeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasarsResponseStatus - -- | The response status code.

dasarsAnalysisScheme - The status of the analysis scheme being deleted.

data DeleteAnalysisSchemeResponse

The result of a DeleteAnalysisScheme request. Contains the status of the deleted analysis scheme.

See: deleteAnalysisSchemeResponse smart constructor.

Response Lenses

dasarsResponseStatus :: Lens' DeleteAnalysisSchemeResponse Int

- | The response status code.

dasarsAnalysisScheme :: Lens' DeleteAnalysisSchemeResponse AnalysisSchemeStatus

The status of the analysis scheme being deleted.