Network.AWS.CloudSearch.DeleteExpression

Description

Removes an Expression from the search domain. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide .

Creating a Request

deleteExpression

Arguments

:: Text
delDomainName

-> Text
delExpressionName

-> DeleteExpression

Creates a value of DeleteExpression with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delDomainName - Undocumented member.

delExpressionName - The name of the Expression to delete.

data DeleteExpression

Container for the parameters to the DeleteExpression operation. Specifies the name of the domain you want to update and the name of the expression you want to delete.

See: deleteExpression smart constructor.

Request Lenses

delDomainName :: Lens' DeleteExpression Text

Undocumented member.

delExpressionName :: Lens' DeleteExpression Text

The name of the Expression to delete.

Destructuring the Response

deleteExpressionResponse

Arguments

:: Int
delrsResponseStatus

-> ExpressionStatus
delrsExpression

-> DeleteExpressionResponse

Creates a value of DeleteExpressionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delrsResponseStatus - -- | The response status code.

delrsExpression - The status of the expression being deleted.

data DeleteExpressionResponse

The result of a DeleteExpression request. Specifies the expression being deleted.

See: deleteExpressionResponse smart constructor.

Response Lenses

delrsResponseStatus :: Lens' DeleteExpressionResponse Int

- | The response status code.

delrsExpression :: Lens' DeleteExpressionResponse ExpressionStatus

The status of the expression being deleted.