Network.AWS.CloudSearch.DeleteIndexField

Description

Removes an IndexField from the search domain. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide .

Creating a Request

deleteIndexField

Arguments

:: Text
difiDomainName

-> Text
difiIndexFieldName

-> DeleteIndexField

Creates a value of DeleteIndexField with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difiDomainName - Undocumented member.

difiIndexFieldName - The name of the index field your want to remove from the domain's indexing options.

data DeleteIndexField

Container for the parameters to the DeleteIndexField operation. Specifies the name of the domain you want to update and the name of the index field you want to delete.

See: deleteIndexField smart constructor.

Request Lenses

difiDomainName :: Lens' DeleteIndexField Text

Undocumented member.

difiIndexFieldName :: Lens' DeleteIndexField Text

The name of the index field your want to remove from the domain's indexing options.

Destructuring the Response

deleteIndexFieldResponse

Arguments

:: Int
difrsResponseStatus

-> IndexFieldStatus
difrsIndexField

-> DeleteIndexFieldResponse

Creates a value of DeleteIndexFieldResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difrsResponseStatus - -- | The response status code.

difrsIndexField - The status of the index field being deleted.

data DeleteIndexFieldResponse

The result of a DeleteIndexField request.

See: deleteIndexFieldResponse smart constructor.

Response Lenses

difrsResponseStatus :: Lens' DeleteIndexFieldResponse Int

- | The response status code.

difrsIndexField :: Lens' DeleteIndexFieldResponse IndexFieldStatus

The status of the index field being deleted.