Network.AWS.CloudSearch.DeleteSuggester

Description

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide .

Creating a Request

deleteSuggester

Arguments

:: Text
ddDomainName

-> Text
ddSuggesterName

-> DeleteSuggester

Creates a value of DeleteSuggester with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddDomainName - Undocumented member.

ddSuggesterName - Specifies the name of the suggester you want to delete.

data DeleteSuggester

Container for the parameters to the DeleteSuggester operation. Specifies the name of the domain you want to update and name of the suggester you want to delete.

See: deleteSuggester smart constructor.

Request Lenses

ddDomainName :: Lens' DeleteSuggester Text

Undocumented member.

ddSuggesterName :: Lens' DeleteSuggester Text

Specifies the name of the suggester you want to delete.

Destructuring the Response

deleteSuggesterResponse

Arguments

:: Int
delersResponseStatus

-> SuggesterStatus
delersSuggester

-> DeleteSuggesterResponse

Creates a value of DeleteSuggesterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delersResponseStatus - -- | The response status code.

delersSuggester - The status of the suggester being deleted.

data DeleteSuggesterResponse

The result of a DeleteSuggester request. Contains the status of the deleted suggester.

See: deleteSuggesterResponse smart constructor.

Response Lenses

delersResponseStatus :: Lens' DeleteSuggesterResponse Int

- | The response status code.

delersSuggester :: Lens' DeleteSuggesterResponse SuggesterStatus

The status of the suggester being deleted.