Network.AWS.CloudSearch.DescribeAvailabilityOptions

Description

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearchDeveloper Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeAvailabilityOptions.html

Request

data DescribeAvailabilityOptions

Request constructor

describeAvailabilityOptions

Arguments

:: Text
daoDomainName

-> DescribeAvailabilityOptions

DescribeAvailabilityOptions constructor.

The fields accessible through corresponding lenses are:

daoDeployed :: Maybe Bool

daoDomainName :: Text

Request lenses

daoDeployed :: Lens' DescribeAvailabilityOptions (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

daoDomainName :: Lens' DescribeAvailabilityOptions Text

The name of the domain you want to describe.

Response

data DescribeAvailabilityOptionsResponse

Response constructor

describeAvailabilityOptionsResponse :: DescribeAvailabilityOptionsResponse

DescribeAvailabilityOptionsResponse constructor.

The fields accessible through corresponding lenses are:

daorAvailabilityOptions :: Maybe AvailabilityOptionsStatus

Response lenses

daorAvailabilityOptions :: Lens' DescribeAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.