Network.AWS.CloudSearch.DescribeExpressions

Description

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide .

Creating a Request

describeExpressions

Arguments

:: Text
deDomainName

-> DescribeExpressions

Creates a value of DescribeExpressions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deDeployed - Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

deExpressionNames - Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

deDomainName - The name of the domain you want to describe.

data DescribeExpressions

Container for the parameters to the DescribeDomains operation. Specifies the name of the domain you want to describe. To restrict the response to particular expressions, specify the names of the expressions you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true .

See: describeExpressions smart constructor.

Request Lenses

deDeployed :: Lens' DescribeExpressions (Maybe Bool)

Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

deExpressionNames :: Lens' DescribeExpressions [Text]

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

deDomainName :: Lens' DescribeExpressions Text

The name of the domain you want to describe.

Destructuring the Response

describeExpressionsResponse

Arguments

:: Int
drsResponseStatus

-> DescribeExpressionsResponse

Creates a value of DescribeExpressionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

drsExpressions - The expressions configured for the domain.

data DescribeExpressionsResponse

The result of a DescribeExpressions request. Contains the expressions configured for the domain specified in the request.

See: describeExpressionsResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DescribeExpressionsResponse Int

- | The response status code.

drsExpressions :: Lens' DescribeExpressionsResponse [ExpressionStatus]

The expressions configured for the domain.