Network.AWS.CloudSearch.DescribeIndexFields

Description

Gets information about the index fields configured for the search domain. Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Domain Information in the Amazon CloudSearch Developer Guide .

Creating a Request

describeIndexFields

Arguments

:: Text
difDomainName

-> DescribeIndexFields

Creates a value of DescribeIndexFields with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difDeployed - Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

difFieldNames - A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

difDomainName - The name of the domain you want to describe.

data DescribeIndexFields

Container for the parameters to the DescribeIndexFields operation. Specifies the name of the domain you want to describe. To restrict the response to particular index fields, specify the names of the index fields you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true .

See: describeIndexFields smart constructor.

Request Lenses

difDeployed :: Lens' DescribeIndexFields (Maybe Bool)

Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

difFieldNames :: Lens' DescribeIndexFields [Text]

A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

difDomainName :: Lens' DescribeIndexFields Text

The name of the domain you want to describe.

Destructuring the Response

describeIndexFieldsResponse

Arguments

:: Int
difsrsResponseStatus

-> DescribeIndexFieldsResponse

Creates a value of DescribeIndexFieldsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difsrsResponseStatus - -- | The response status code.

difsrsIndexFields - The index fields configured for the domain.

data DescribeIndexFieldsResponse

The result of a DescribeIndexFields request. Contains the index fields configured for the domain specified in the request.

See: describeIndexFieldsResponse smart constructor.

Response Lenses

difsrsResponseStatus :: Lens' DescribeIndexFieldsResponse Int

- | The response status code.

difsrsIndexFields :: Lens' DescribeIndexFieldsResponse [IndexFieldStatus]

The index fields configured for the domain.