Network.AWS.CloudSearch.DescribeServiceAccessPolicies

Description

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide .

Creating a Request

describeServiceAccessPolicies

Arguments

:: Text
dsapDomainName

-> DescribeServiceAccessPolicies

Creates a value of DescribeServiceAccessPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsapDeployed - Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

dsapDomainName - The name of the domain you want to describe.

data DescribeServiceAccessPolicies

Container for the parameters to the DescribeServiceAccessPolicies operation. Specifies the name of the domain you want to describe. To show the active configuration and exclude any pending changes, set the Deployed option to true .

See: describeServiceAccessPolicies smart constructor.

Request Lenses

dsapDeployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

Whether to display the deployed configuration (true ) or include any pending changes (false ). Defaults to false .

dsapDomainName :: Lens' DescribeServiceAccessPolicies Text

The name of the domain you want to describe.

Destructuring the Response

describeServiceAccessPoliciesResponse

Arguments

:: Int
dsaprsResponseStatus

-> AccessPoliciesStatus
dsaprsAccessPolicies

-> DescribeServiceAccessPoliciesResponse

Creates a value of DescribeServiceAccessPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsaprsResponseStatus - -- | The response status code.

dsaprsAccessPolicies - The access rules configured for the domain specified in the request.

data DescribeServiceAccessPoliciesResponse

The result of a DescribeServiceAccessPolicies request.

See: describeServiceAccessPoliciesResponse smart constructor.

Response Lenses

dsaprsResponseStatus :: Lens' DescribeServiceAccessPoliciesResponse Int

- | The response status code.

dsaprsAccessPolicies :: Lens' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain specified in the request.