Network.AWS.CloudSearch.DescribeServiceAccessPolicies

Description

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

http://docs.aws.amazon.com/cloudsearch/latest/developerguide/API_DescribeServiceAccessPolicies.html

Request

data DescribeServiceAccessPolicies

Request constructor

describeServiceAccessPolicies

Arguments

:: Text
dsapDomainName

-> DescribeServiceAccessPolicies

DescribeServiceAccessPolicies constructor.

The fields accessible through corresponding lenses are:

dsapDeployed :: Maybe Bool

dsapDomainName :: Text

Request lenses

dsapDeployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

dsapDomainName :: Lens' DescribeServiceAccessPolicies Text

The name of the domain you want to describe.

Response

data DescribeServiceAccessPoliciesResponse

Response constructor

describeServiceAccessPoliciesResponse

Arguments

:: AccessPoliciesStatus
dsaprAccessPolicies

-> DescribeServiceAccessPoliciesResponse

DescribeServiceAccessPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

dsaprAccessPolicies :: AccessPoliciesStatus

Response lenses

dsaprAccessPolicies :: Lens' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain specified in the request.