Network.AWS.CloudSearch.IndexDocuments

Description

Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments .

Creating a Request

indexDocuments

Arguments

:: Text
idDomainName

-> IndexDocuments

Creates a value of IndexDocuments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idDomainName - Undocumented member.

data IndexDocuments

Container for the parameters to the IndexDocuments operation. Specifies the name of the domain you want to re-index.

See: indexDocuments smart constructor.

Request Lenses

idDomainName :: Lens' IndexDocuments Text

Undocumented member.

Destructuring the Response

indexDocumentsResponse

Arguments

:: Int
idrsResponseStatus

-> IndexDocumentsResponse

Creates a value of IndexDocumentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idrsFieldNames - The names of the fields that are currently being indexed.

idrsResponseStatus - -- | The response status code.

data IndexDocumentsResponse

The result of an IndexDocuments request. Contains the status of the indexing operation, including the fields being indexed.

See: indexDocumentsResponse smart constructor.

Response Lenses

idrsFieldNames :: Lens' IndexDocumentsResponse [Text]

The names of the fields that are currently being indexed.

idrsResponseStatus :: Lens' IndexDocumentsResponse Int

- | The response status code.