Network.AWS.CloudSearch.Types

Description

Service Configuration

cloudSearch :: Service

API version 2013-01-01 of the Amazon CloudSearch SDK configuration.

Errors

_BaseException :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred while processing the request.

_DisabledOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it attempted an operation which is not enabled.

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred while processing the request. If this problem persists, report an issue from the Service Health Dashboard .

_InvalidTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it specified an invalid type definition.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it attempted to reference a resource that does not exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because a resource limit has already been met.

AlgorithmicStemming

data AlgorithmicStemming

Constructors

ASFull

ASLight

ASMinimal

ASNone

AnalysisSchemeLanguage

data AnalysisSchemeLanguage

An IETF RFC 4646 language code or mul for multiple languages.

Constructors

AR

BG

CA

CS

DA

DE

EL

EN

ES

EU

FA

FI

FR

GA

GL

HE

HI

HU

HY

IT

Id

JA

KO

LV

Mul

NL

NO

PT

RO

RU

SV

TH

TR

ZhHans

ZhHant

IndexFieldType

data IndexFieldType

The type of field. The valid options for a field depend on the field type. For more information about the supported field types, see Configuring Index Fields in the Amazon CloudSearch Developer Guide .

Constructors

Date

DateArray

Double

DoubleArray

Int

IntArray

Latlon

Literal

LiteralArray

Text

TextArray

OptionState

data OptionState

The state of processing a change to an option. One of:

RequiresIndexDocuments: The option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. * Processing: The option's latest value is in the process of being activated. * Active: The option's latest value is fully deployed. * FailedToValidate: The option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

Constructors

Active

FailedToValidate

Processing

RequiresIndexDocuments

PartitionInstanceType

data PartitionInstanceType

The instance type (such as search.m1.small ) on which an index partition is hosted.

Constructors

Search_M1_Large

Search_M1_Small

Search_M2_2XLarge

Search_M2_XLarge

Search_M3_2XLarge

Search_M3_Large

Search_M3_Medium

Search_M3_XLarge

SuggesterFuzzyMatching

data SuggesterFuzzyMatching

Constructors

High

Low

None

AccessPoliciesStatus

data AccessPoliciesStatus

The configured access rules for the domain's document and search endpoints, and the current status of those rules.

See: accessPoliciesStatus smart constructor.

accessPoliciesStatus

Arguments

:: Text
apsOptions

-> OptionStatus
apsStatus

-> AccessPoliciesStatus

Creates a value of AccessPoliciesStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apsOptions - Undocumented member.

apsStatus - Undocumented member.

apsOptions :: Lens' AccessPoliciesStatus Text

Undocumented member.

apsStatus :: Lens' AccessPoliciesStatus OptionStatus

Undocumented member.

AnalysisOptions

data AnalysisOptions

Synonyms, stopwords, and stemming options for an analysis scheme. Includes tokenization dictionary for Japanese.

See: analysisOptions smart constructor.

analysisOptions :: AnalysisOptions

Creates a value of AnalysisOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aoAlgorithmicStemming - The level of algorithmic stemming to perform: none , minimal , light , or full . The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearch Developer Guide

aoStopwords - A JSON array of terms to ignore during indexing and searching. For example, ["a", "an", "the", "of"] . The stopwords dictionary must explicitly list each word you want to ignore. Wildcards and regular expressions are not supported.

aoJapaneseTokenizationDictionary - A JSON array that contains a collection of terms, tokens, readings and part of speech for Japanese Tokenizaiton. The Japanese tokenization dictionary enables you to override the default tokenization for selected terms. This is only valid for Japanese language fields.

aoSynonyms - A JSON object that defines synonym groups and aliases. A synonym group is an array of arrays, where each sub-array is a group of terms where each term in the group is considered a synonym of every other term in the group. The aliases value is an object that contains a collection of string:value pairs where the string specifies a term and the array of values specifies each of the aliases for that term. An alias is considered a synonym of the specified term, but the term is not considered a synonym of the alias. For more information about specifying synonyms, see Synonyms in the Amazon CloudSearch Developer Guide .

aoStemmingDictionary - A JSON object that contains a collection of string:value pairs that each map a term to its stem. For example, {"term1": "stem1", "term2": "stem2", "term3": "stem3"} . The stemming dictionary is applied in addition to any algorithmic stemming. This enables you to override the results of the algorithmic stemming to correct specific cases of overstemming or understemming. The maximum size of a stemming dictionary is 500 KB.

aoAlgorithmicStemming :: Lens' AnalysisOptions (Maybe AlgorithmicStemming)

The level of algorithmic stemming to perform: none , minimal , light , or full . The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearch Developer Guide

aoStopwords :: Lens' AnalysisOptions (Maybe Text)

A JSON array of terms to ignore during indexing and searching. For example, ["a", "an", "the", "of"] . The stopwords dictionary must explicitly list each word you want to ignore. Wildcards and regular expressions are not supported.

aoJapaneseTokenizationDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON array that contains a collection of terms, tokens, readings and part of speech for Japanese Tokenizaiton. The Japanese tokenization dictionary enables you to override the default tokenization for selected terms. This is only valid for Japanese language fields.

aoSynonyms :: Lens' AnalysisOptions (Maybe Text)

A JSON object that defines synonym groups and aliases. A synonym group is an array of arrays, where each sub-array is a group of terms where each term in the group is considered a synonym of every other term in the group. The aliases value is an object that contains a collection of string:value pairs where the string specifies a term and the array of values specifies each of the aliases for that term. An alias is considered a synonym of the specified term, but the term is not considered a synonym of the alias. For more information about specifying synonyms, see Synonyms in the Amazon CloudSearch Developer Guide .

aoStemmingDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON object that contains a collection of string:value pairs that each map a term to its stem. For example, {"term1": "stem1", "term2": "stem2", "term3": "stem3"} . The stemming dictionary is applied in addition to any algorithmic stemming. This enables you to override the results of the algorithmic stemming to correct specific cases of overstemming or understemming. The maximum size of a stemming dictionary is 500 KB.

AnalysisScheme

data AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms , Stopwords , StemmingDictionary , JapaneseTokenizationDictionary and AlgorithmicStemming .

See: analysisScheme smart constructor.

analysisScheme

Arguments

:: Text
asAnalysisSchemeName

-> AnalysisSchemeLanguage
asAnalysisSchemeLanguage

-> AnalysisScheme

Creates a value of AnalysisScheme with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asAnalysisOptions - Undocumented member.

asAnalysisSchemeName - Undocumented member.

asAnalysisSchemeLanguage - Undocumented member.

asAnalysisOptions :: Lens' AnalysisScheme (Maybe AnalysisOptions)

Undocumented member.

asAnalysisSchemeName :: Lens' AnalysisScheme Text

Undocumented member.

asAnalysisSchemeLanguage :: Lens' AnalysisScheme AnalysisSchemeLanguage

Undocumented member.

AnalysisSchemeStatus

data AnalysisSchemeStatus

The status and configuration of an AnalysisScheme .

See: analysisSchemeStatus smart constructor.

analysisSchemeStatus

Arguments

:: AnalysisScheme
assOptions

-> OptionStatus
assStatus

-> AnalysisSchemeStatus

Creates a value of AnalysisSchemeStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

assOptions - Undocumented member.

assStatus - Undocumented member.

assOptions :: Lens' AnalysisSchemeStatus AnalysisScheme

Undocumented member.

assStatus :: Lens' AnalysisSchemeStatus OptionStatus

Undocumented member.

AvailabilityOptionsStatus

data AvailabilityOptionsStatus

The status and configuration of the domain's availability options.

See: availabilityOptionsStatus smart constructor.

availabilityOptionsStatus

Arguments

:: Bool
aosOptions

-> OptionStatus
aosStatus

-> AvailabilityOptionsStatus

Creates a value of AvailabilityOptionsStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aosOptions - The availability options configured for the domain.

aosStatus - Undocumented member.

aosOptions :: Lens' AvailabilityOptionsStatus Bool

The availability options configured for the domain.

aosStatus :: Lens' AvailabilityOptionsStatus OptionStatus

Undocumented member.

DateArrayOptions

data DateArrayOptions

Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array . All options are enabled by default.

See: dateArrayOptions smart constructor.

dateArrayOptions :: DateArrayOptions

Creates a value of DateArrayOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daosSourceFields - A list of source fields to map to the field.

daosReturnEnabled - Whether the contents of the field can be returned in the search results.

daosFacetEnabled - Whether facet information can be returned for the field.

daosSearchEnabled - Whether the contents of the field are searchable.

daosDefaultValue - A value to use for the field if the field isn't specified for a document.

daosSourceFields :: Lens' DateArrayOptions (Maybe Text)

A list of source fields to map to the field.

daosReturnEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

daosFacetEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

daosSearchEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

daosDefaultValue :: Lens' DateArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

DateOptions

data DateOptions

Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date . All options are enabled by default.

See: dateOptions smart constructor.

dateOptions :: DateOptions

Creates a value of DateOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

doSourceField - Undocumented member.

doReturnEnabled - Whether the contents of the field can be returned in the search results.

doFacetEnabled - Whether facet information can be returned for the field.

doSearchEnabled - Whether the contents of the field are searchable.

doSortEnabled - Whether the field can be used to sort the search results.

doDefaultValue - A value to use for the field if the field isn't specified for a document.

doSourceField :: Lens' DateOptions (Maybe Text)

Undocumented member.

doReturnEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

doFacetEnabled :: Lens' DateOptions (Maybe Bool)

Whether facet information can be returned for the field.

doSearchEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field are searchable.

doSortEnabled :: Lens' DateOptions (Maybe Bool)

Whether the field can be used to sort the search results.

doDefaultValue :: Lens' DateOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

DocumentSuggesterOptions

data DocumentSuggesterOptions

Options for a search suggester.

See: documentSuggesterOptions smart constructor.

documentSuggesterOptions

Arguments

:: Text
dsoSourceField

-> DocumentSuggesterOptions

Creates a value of DocumentSuggesterOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsoSortExpression - An expression that computes a score for each suggestion to control how they are sorted. The scores are rounded to the nearest integer, with a floor of 0 and a ceiling of 2^31-1. A document's relevance score is not computed for suggestions, so sort expressions cannot reference the _score value. To sort suggestions using a numeric field or existing expression, simply specify the name of the field or expression. If no expression is configured for the suggester, the suggestions are sorted with the closest matches listed first.

dsoFuzzyMatching - The level of fuzziness allowed when suggesting matches for a string: none , low , or high . With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

dsoSourceField - The name of the index field you want to use for suggestions.

dsoSortExpression :: Lens' DocumentSuggesterOptions (Maybe Text)

An expression that computes a score for each suggestion to control how they are sorted. The scores are rounded to the nearest integer, with a floor of 0 and a ceiling of 2^31-1. A document's relevance score is not computed for suggestions, so sort expressions cannot reference the _score value. To sort suggestions using a numeric field or existing expression, simply specify the name of the field or expression. If no expression is configured for the suggester, the suggestions are sorted with the closest matches listed first.

dsoFuzzyMatching :: Lens' DocumentSuggesterOptions (Maybe SuggesterFuzzyMatching)

The level of fuzziness allowed when suggesting matches for a string: none , low , or high . With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

dsoSourceField :: Lens' DocumentSuggesterOptions Text

The name of the index field you want to use for suggestions.

DomainStatus

data DomainStatus

The current status of the search domain.

See: domainStatus smart constructor.

domainStatus

Arguments

:: Text
dsDomainId

-> Text
dsDomainName

-> Bool
dsRequiresIndexDocuments

-> DomainStatus

Creates a value of DomainStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsSearchInstanceCount - The number of search instances that are available to process search requests.

dsSearchInstanceType - The instance type that is being used to process search requests.

dsDocService - The service endpoint for updating documents in a search domain.

dsARN - Undocumented member.

dsCreated - True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

dsSearchService - The service endpoint for requesting search results from a search domain.

dsLimits - Undocumented member.

dsSearchPartitionCount - The number of partitions across which the search index is spread.

dsDeleted - True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

dsProcessing - True if processing is being done to activate the current domain configuration.

dsDomainId - Undocumented member.

dsDomainName - Undocumented member.

dsRequiresIndexDocuments - True if IndexDocuments needs to be called to activate the current domain configuration.

dsSearchInstanceCount :: Lens' DomainStatus (Maybe Natural)

The number of search instances that are available to process search requests.

dsSearchInstanceType :: Lens' DomainStatus (Maybe Text)

The instance type that is being used to process search requests.

dsDocService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for updating documents in a search domain.

dsARN :: Lens' DomainStatus (Maybe Text)

Undocumented member.

dsCreated :: Lens' DomainStatus (Maybe Bool)

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

dsSearchService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for requesting search results from a search domain.

dsLimits :: Lens' DomainStatus (Maybe Limits)

Undocumented member.

dsSearchPartitionCount :: Lens' DomainStatus (Maybe Natural)

The number of partitions across which the search index is spread.

dsDeleted :: Lens' DomainStatus (Maybe Bool)

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

dsProcessing :: Lens' DomainStatus (Maybe Bool)

True if processing is being done to activate the current domain configuration.

dsDomainId :: Lens' DomainStatus Text

Undocumented member.

dsDomainName :: Lens' DomainStatus Text

Undocumented member.

dsRequiresIndexDocuments :: Lens' DomainStatus Bool

True if IndexDocuments needs to be called to activate the current domain configuration.

DoubleArrayOptions

data DoubleArrayOptions

Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array . All options are enabled by default.

See: doubleArrayOptions smart constructor.

doubleArrayOptions :: DoubleArrayOptions

Creates a value of DoubleArrayOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daoSourceFields - A list of source fields to map to the field.

daoReturnEnabled - Whether the contents of the field can be returned in the search results.

daoFacetEnabled - Whether facet information can be returned for the field.

daoSearchEnabled - Whether the contents of the field are searchable.

daoDefaultValue - A value to use for the field if the field isn't specified for a document.

daoSourceFields :: Lens' DoubleArrayOptions (Maybe Text)

A list of source fields to map to the field.

daoReturnEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

daoFacetEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

daoSearchEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

daoDefaultValue :: Lens' DoubleArrayOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document.

DoubleOptions

data DoubleOptions

Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double . All options are enabled by default.

See: doubleOptions smart constructor.

doubleOptions :: DoubleOptions

Creates a value of DoubleOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dSourceField - The name of the source field to map to the field.

dReturnEnabled - Whether the contents of the field can be returned in the search results.

dFacetEnabled - Whether facet information can be returned for the field.

dSearchEnabled - Whether the contents of the field are searchable.

dSortEnabled - Whether the field can be used to sort the search results.

dDefaultValue - A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

dSourceField :: Lens' DoubleOptions (Maybe Text)

The name of the source field to map to the field.

dReturnEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

dFacetEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether facet information can be returned for the field.

dSearchEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field are searchable.

dSortEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the field can be used to sort the search results.

dDefaultValue :: Lens' DoubleOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

Expression

data Expression

A named expression that can be evaluated at search time. Can be used to sort the search results, define other expressions, or return computed information in the search results.

See: expression smart constructor.

expression

Arguments

:: Text
eExpressionName

-> Text
eExpressionValue

-> Expression

Creates a value of Expression with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eExpressionName - Undocumented member.

eExpressionValue - Undocumented member.

eExpressionName :: Lens' Expression Text

Undocumented member.

eExpressionValue :: Lens' Expression Text

Undocumented member.

ExpressionStatus

data ExpressionStatus

The value of an Expression and its current status.

See: expressionStatus smart constructor.

expressionStatus

Arguments

:: Expression
esOptions

-> OptionStatus
esStatus

-> ExpressionStatus

Creates a value of ExpressionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esOptions - The expression that is evaluated for sorting while processing a search request.

esStatus - Undocumented member.

esOptions :: Lens' ExpressionStatus Expression

The expression that is evaluated for sorting while processing a search request.

esStatus :: Lens' ExpressionStatus OptionStatus

Undocumented member.

IndexField

data IndexField

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType .

See: indexField smart constructor.

indexField

Arguments

:: Text
ifIndexFieldName

-> IndexFieldType
ifIndexFieldType

-> IndexField

Creates a value of IndexField with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ifDoubleArrayOptions - Undocumented member.

ifDateOptions - Undocumented member.

ifTextArrayOptions - Undocumented member.

ifDoubleOptions - Undocumented member.

ifTextOptions - Undocumented member.

ifLatLonOptions - Undocumented member.

ifLiteralArrayOptions - Undocumented member.

ifIntArrayOptions - Undocumented member.

ifDateArrayOptions - Undocumented member.

ifIntOptions - Undocumented member.

ifLiteralOptions - Undocumented member.

ifIndexFieldName - A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options. Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported. The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id .

ifIndexFieldType - Undocumented member.

ifDoubleArrayOptions :: Lens' IndexField (Maybe DoubleArrayOptions)

Undocumented member.

ifDateOptions :: Lens' IndexField (Maybe DateOptions)

Undocumented member.

ifTextArrayOptions :: Lens' IndexField (Maybe TextArrayOptions)

Undocumented member.

ifDoubleOptions :: Lens' IndexField (Maybe DoubleOptions)

Undocumented member.

ifTextOptions :: Lens' IndexField (Maybe TextOptions)

Undocumented member.

ifLatLonOptions :: Lens' IndexField (Maybe LatLonOptions)

Undocumented member.

ifLiteralArrayOptions :: Lens' IndexField (Maybe LiteralArrayOptions)

Undocumented member.

ifIntArrayOptions :: Lens' IndexField (Maybe IntArrayOptions)

Undocumented member.

ifDateArrayOptions :: Lens' IndexField (Maybe DateArrayOptions)

Undocumented member.

ifIntOptions :: Lens' IndexField (Maybe IntOptions)

Undocumented member.

ifLiteralOptions :: Lens' IndexField (Maybe LiteralOptions)

Undocumented member.

ifIndexFieldName :: Lens' IndexField Text

A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options. Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported. The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id .

ifIndexFieldType :: Lens' IndexField IndexFieldType

Undocumented member.

IndexFieldStatus

data IndexFieldStatus

The value of an IndexField and its current status.

See: indexFieldStatus smart constructor.

indexFieldStatus

Arguments

:: IndexField
ifsOptions

-> OptionStatus
ifsStatus

-> IndexFieldStatus

Creates a value of IndexFieldStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ifsOptions - Undocumented member.

ifsStatus - Undocumented member.

ifsOptions :: Lens' IndexFieldStatus IndexField

Undocumented member.

ifsStatus :: Lens' IndexFieldStatus OptionStatus

Undocumented member.

IntArrayOptions

data IntArrayOptions

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array . All options are enabled by default.

See: intArrayOptions smart constructor.

intArrayOptions :: IntArrayOptions

Creates a value of IntArrayOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iaoSourceFields - A list of source fields to map to the field.

iaoReturnEnabled - Whether the contents of the field can be returned in the search results.

iaoFacetEnabled - Whether facet information can be returned for the field.

iaoSearchEnabled - Whether the contents of the field are searchable.

iaoDefaultValue - A value to use for the field if the field isn't specified for a document.

iaoSourceFields :: Lens' IntArrayOptions (Maybe Text)

A list of source fields to map to the field.

iaoReturnEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

iaoFacetEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

iaoSearchEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

iaoDefaultValue :: Lens' IntArrayOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document.

IntOptions

data IntOptions

Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int . All options are enabled by default.

See: intOptions smart constructor.

intOptions :: IntOptions

Creates a value of IntOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ioSourceField - The name of the source field to map to the field.

ioReturnEnabled - Whether the contents of the field can be returned in the search results.

ioFacetEnabled - Whether facet information can be returned for the field.

ioSearchEnabled - Whether the contents of the field are searchable.

ioSortEnabled - Whether the field can be used to sort the search results.

ioDefaultValue - A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

ioSourceField :: Lens' IntOptions (Maybe Text)

The name of the source field to map to the field.

ioReturnEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

ioFacetEnabled :: Lens' IntOptions (Maybe Bool)

Whether facet information can be returned for the field.

ioSearchEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field are searchable.

ioSortEnabled :: Lens' IntOptions (Maybe Bool)

Whether the field can be used to sort the search results.

ioDefaultValue :: Lens' IntOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

LatLonOptions

data LatLonOptions

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon . All options are enabled by default.

See: latLonOptions smart constructor.

latLonOptions :: LatLonOptions

Creates a value of LatLonOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lloSourceField - Undocumented member.

lloReturnEnabled - Whether the contents of the field can be returned in the search results.

lloFacetEnabled - Whether facet information can be returned for the field.

lloSearchEnabled - Whether the contents of the field are searchable.

lloSortEnabled - Whether the field can be used to sort the search results.

lloDefaultValue - A value to use for the field if the field isn't specified for a document.

lloSourceField :: Lens' LatLonOptions (Maybe Text)

Undocumented member.

lloReturnEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

lloFacetEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether facet information can be returned for the field.

lloSearchEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field are searchable.

lloSortEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the field can be used to sort the search results.

lloDefaultValue :: Lens' LatLonOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

Limits

data Limits

See: limits smart constructor.

limits

Arguments

:: Natural
lMaximumReplicationCount

-> Natural
lMaximumPartitionCount

-> Limits

Creates a value of Limits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lMaximumReplicationCount - Undocumented member.

lMaximumPartitionCount - Undocumented member.

lMaximumReplicationCount :: Lens' Limits Natural

Undocumented member.

lMaximumPartitionCount :: Lens' Limits Natural

Undocumented member.

LiteralArrayOptions

data LiteralArrayOptions

Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array . All options are enabled by default.

See: literalArrayOptions smart constructor.

literalArrayOptions :: LiteralArrayOptions

Creates a value of LiteralArrayOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laoSourceFields - A list of source fields to map to the field.

laoReturnEnabled - Whether the contents of the field can be returned in the search results.

laoFacetEnabled - Whether facet information can be returned for the field.

laoSearchEnabled - Whether the contents of the field are searchable.

laoDefaultValue - A value to use for the field if the field isn't specified for a document.

laoSourceFields :: Lens' LiteralArrayOptions (Maybe Text)

A list of source fields to map to the field.

laoReturnEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

laoFacetEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

laoSearchEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

laoDefaultValue :: Lens' LiteralArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

LiteralOptions

data LiteralOptions

Options for literal field. Present if IndexFieldType specifies the field is of type literal . All options are enabled by default.

See: literalOptions smart constructor.

literalOptions :: LiteralOptions

Creates a value of LiteralOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loSourceField - Undocumented member.

loReturnEnabled - Whether the contents of the field can be returned in the search results.

loFacetEnabled - Whether facet information can be returned for the field.

loSearchEnabled - Whether the contents of the field are searchable.

loSortEnabled - Whether the field can be used to sort the search results.

loDefaultValue - A value to use for the field if the field isn't specified for a document.

loSourceField :: Lens' LiteralOptions (Maybe Text)

Undocumented member.

loReturnEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

loFacetEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether facet information can be returned for the field.

loSearchEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field are searchable.

loSortEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the field can be used to sort the search results.

loDefaultValue :: Lens' LiteralOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

OptionStatus

data OptionStatus

The status of domain configuration option.

See: optionStatus smart constructor.

optionStatus

Arguments

:: UTCTime
osCreationDate

-> UTCTime
osUpdateDate

-> OptionState
osState

-> OptionStatus

Creates a value of OptionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

osPendingDeletion - Indicates that the option will be deleted once processing is complete.

osUpdateVersion - A unique integer that indicates when this option was last updated.

osCreationDate - A timestamp for when this option was created.

osUpdateDate - A timestamp for when this option was last updated.

osState - The state of processing a change to an option. Possible values: * RequiresIndexDocuments : the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. * Processing : the option's latest value is in the process of being activated. * Active : the option's latest value is completely deployed. * FailedToValidate : the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

osPendingDeletion :: Lens' OptionStatus (Maybe Bool)

Indicates that the option will be deleted once processing is complete.

osUpdateVersion :: Lens' OptionStatus (Maybe Natural)

A unique integer that indicates when this option was last updated.

osCreationDate :: Lens' OptionStatus UTCTime

A timestamp for when this option was created.

osUpdateDate :: Lens' OptionStatus UTCTime

A timestamp for when this option was last updated.

osState :: Lens' OptionStatus OptionState

The state of processing a change to an option. Possible values: * RequiresIndexDocuments : the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. * Processing : the option's latest value is in the process of being activated. * Active : the option's latest value is completely deployed. * FailedToValidate : the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

ScalingParameters

data ScalingParameters

The desired instance type and desired number of replicas of each index partition.

See: scalingParameters smart constructor.

scalingParameters :: ScalingParameters

Creates a value of ScalingParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spDesiredInstanceType - The instance type that you want to preconfigure for your domain. For example, search.m1.small .

spDesiredReplicationCount - The number of replicas you want to preconfigure for each index partition.

spDesiredPartitionCount - The number of partitions you want to preconfigure for your domain. Only valid when you select m2.2xlarge as the desired instance type.

spDesiredInstanceType :: Lens' ScalingParameters (Maybe PartitionInstanceType)

The instance type that you want to preconfigure for your domain. For example, search.m1.small .

spDesiredReplicationCount :: Lens' ScalingParameters (Maybe Natural)

The number of replicas you want to preconfigure for each index partition.

spDesiredPartitionCount :: Lens' ScalingParameters (Maybe Natural)

The number of partitions you want to preconfigure for your domain. Only valid when you select m2.2xlarge as the desired instance type.

ScalingParametersStatus

data ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

See: scalingParametersStatus smart constructor.

scalingParametersStatus

Arguments

:: ScalingParameters
spsOptions

-> OptionStatus
spsStatus

-> ScalingParametersStatus

Creates a value of ScalingParametersStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spsOptions - Undocumented member.

spsStatus - Undocumented member.

spsOptions :: Lens' ScalingParametersStatus ScalingParameters

Undocumented member.

spsStatus :: Lens' ScalingParametersStatus OptionStatus

Undocumented member.

ServiceEndpoint

data ServiceEndpoint

The endpoint to which service requests can be submitted.

See: serviceEndpoint smart constructor.

serviceEndpoint :: ServiceEndpoint

Creates a value of ServiceEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

seEndpoint - Undocumented member.

seEndpoint :: Lens' ServiceEndpoint (Maybe Text)

Undocumented member.

Suggester

data Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching , SortExpression .

See: suggester smart constructor.

suggester

Arguments

:: Text
sSuggesterName

-> DocumentSuggesterOptions
sDocumentSuggesterOptions

-> Suggester

Creates a value of Suggester with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sSuggesterName - Undocumented member.

sDocumentSuggesterOptions - Undocumented member.

sSuggesterName :: Lens' Suggester Text

Undocumented member.

sDocumentSuggesterOptions :: Lens' Suggester DocumentSuggesterOptions

Undocumented member.

SuggesterStatus

data SuggesterStatus

The value of a Suggester and its current status.

See: suggesterStatus smart constructor.

suggesterStatus

Arguments

:: Suggester
ssOptions

-> OptionStatus
ssStatus

-> SuggesterStatus

Creates a value of SuggesterStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssOptions - Undocumented member.

ssStatus - Undocumented member.

ssOptions :: Lens' SuggesterStatus Suggester

Undocumented member.

ssStatus :: Lens' SuggesterStatus OptionStatus

Undocumented member.

TextArrayOptions

data TextArrayOptions

Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array . A text-array field is always searchable. All options are enabled by default.

See: textArrayOptions smart constructor.

textArrayOptions :: TextArrayOptions

Creates a value of TextArrayOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

taoSourceFields - A list of source fields to map to the field.

taoReturnEnabled - Whether the contents of the field can be returned in the search results.

taoAnalysisScheme - The name of an analysis scheme for a text-array field.

taoHighlightEnabled - Whether highlights can be returned for the field.

taoDefaultValue - A value to use for the field if the field isn't specified for a document.

taoSourceFields :: Lens' TextArrayOptions (Maybe Text)

A list of source fields to map to the field.

taoReturnEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

taoAnalysisScheme :: Lens' TextArrayOptions (Maybe Text)

The name of an analysis scheme for a text-array field.

taoHighlightEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether highlights can be returned for the field.

taoDefaultValue :: Lens' TextArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

TextOptions

data TextOptions

Options for text field. Present if IndexFieldType specifies the field is of type text . A text field is always searchable. All options are enabled by default.

See: textOptions smart constructor.

textOptions :: TextOptions

Creates a value of TextOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

toSourceField - Undocumented member.

toReturnEnabled - Whether the contents of the field can be returned in the search results.

toAnalysisScheme - The name of an analysis scheme for a text field.

toHighlightEnabled - Whether highlights can be returned for the field.

toSortEnabled - Whether the field can be used to sort the search results.

toDefaultValue - A value to use for the field if the field isn't specified for a document.

toSourceField :: Lens' TextOptions (Maybe Text)

Undocumented member.

toReturnEnabled :: Lens' TextOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

toAnalysisScheme :: Lens' TextOptions (Maybe Text)

The name of an analysis scheme for a text field.

toHighlightEnabled :: Lens' TextOptions (Maybe Bool)

Whether highlights can be returned for the field.

toSortEnabled :: Lens' TextOptions (Maybe Bool)

Whether the field can be used to sort the search results.

toDefaultValue :: Lens' TextOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.