Network.AWS.CloudSearch.Types

Service

data CloudSearch

Version 2013-01-01 of the Amazon CloudSearch service.

Error

data RESTError :: *

XML

ns :: Text

DomainStatus

data DomainStatus

domainStatus

Arguments

:: Text
dsDomainId

-> Text
dsDomainName

-> Bool
dsRequiresIndexDocuments

-> DomainStatus

DomainStatus constructor.

The fields accessible through corresponding lenses are:

dsARN :: Maybe Text

dsCreated :: Maybe Bool

dsDeleted :: Maybe Bool

dsDocService :: Maybe ServiceEndpoint

dsDomainId :: Text

dsDomainName :: Text

dsLimits :: Maybe Limits

dsProcessing :: Maybe Bool

dsRequiresIndexDocuments :: Bool

dsSearchInstanceCount :: Maybe Natural

dsSearchInstanceType :: Maybe Text

dsSearchPartitionCount :: Maybe Natural

dsSearchService :: Maybe ServiceEndpoint

dsARN :: Lens' DomainStatus (Maybe Text)

dsCreated :: Lens' DomainStatus (Maybe Bool)

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

dsDeleted :: Lens' DomainStatus (Maybe Bool)

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

dsDocService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for updating documents in a search domain.

dsDomainId :: Lens' DomainStatus Text

dsDomainName :: Lens' DomainStatus Text

dsLimits :: Lens' DomainStatus (Maybe Limits)

dsProcessing :: Lens' DomainStatus (Maybe Bool)

True if processing is being done to activate the current domain configuration.

dsRequiresIndexDocuments :: Lens' DomainStatus Bool

True if IndexDocuments needs to be called to activate the current domain configuration.

dsSearchInstanceCount :: Lens' DomainStatus (Maybe Natural)

The number of search instances that are available to process search requests.

dsSearchInstanceType :: Lens' DomainStatus (Maybe Text)

The instance type that is being used to process search requests.

dsSearchPartitionCount :: Lens' DomainStatus (Maybe Natural)

The number of partitions across which the search index is spread.

dsSearchService :: Lens' DomainStatus (Maybe ServiceEndpoint)

The service endpoint for requesting search results from a search domain.

DocumentSuggesterOptions

data DocumentSuggesterOptions

documentSuggesterOptions

Arguments

:: Text
dsoSourceField

-> DocumentSuggesterOptions

DocumentSuggesterOptions constructor.

The fields accessible through corresponding lenses are:

dsoFuzzyMatching :: Maybe SuggesterFuzzyMatching

dsoSortExpression :: Maybe Text

dsoSourceField :: Text

dsoFuzzyMatching :: Lens' DocumentSuggesterOptions (Maybe SuggesterFuzzyMatching)

The level of fuzziness allowed when suggesting matches for a string: none, low , or high. With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

dsoSortExpression :: Lens' DocumentSuggesterOptions (Maybe Text)

An expression that computes a score for each suggestion to control how they are sorted. The scores are rounded to the nearest integer, with a floor of 0 and a ceiling of 2^31-1. A document's relevance score is not computed for suggestions, so sort expressions cannot reference the _score value. To sort suggestions using a numeric field or existing expression, simply specify the name of the field or expression. If no expression is configured for the suggester, the suggestions are sorted with the closest matches listed first.

dsoSourceField :: Lens' DocumentSuggesterOptions Text

The name of the index field you want to use for suggestions.

DoubleArrayOptions

data DoubleArrayOptions

doubleArrayOptions :: DoubleArrayOptions

DoubleArrayOptions constructor.

The fields accessible through corresponding lenses are:

daoDefaultValue :: Maybe Double

daoFacetEnabled :: Maybe Bool

daoReturnEnabled :: Maybe Bool

daoSearchEnabled :: Maybe Bool

daoSourceFields :: Maybe Text

daoDefaultValue :: Lens' DoubleArrayOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document.

daoFacetEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

daoReturnEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

daoSearchEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

daoSourceFields :: Lens' DoubleArrayOptions (Maybe Text)

A list of source fields to map to the field.

IndexField

data IndexField

indexField

Arguments

:: Text
ifIndexFieldName

-> IndexFieldType
ifIndexFieldType

-> IndexField

IndexField constructor.

The fields accessible through corresponding lenses are:

ifDateArrayOptions :: Maybe DateArrayOptions

ifDateOptions :: Maybe DateOptions

ifDoubleArrayOptions :: Maybe DoubleArrayOptions

ifDoubleOptions :: Maybe DoubleOptions

ifIndexFieldName :: Text

ifIndexFieldType :: IndexFieldType

ifIntArrayOptions :: Maybe IntArrayOptions

ifIntOptions :: Maybe IntOptions

ifLatLonOptions :: Maybe LatLonOptions

ifLiteralArrayOptions :: Maybe LiteralArrayOptions

ifLiteralOptions :: Maybe LiteralOptions

ifTextArrayOptions :: Maybe TextArrayOptions

ifTextOptions :: Maybe TextOptions

ifDateArrayOptions :: Lens' IndexField (Maybe DateArrayOptions)

ifDateOptions :: Lens' IndexField (Maybe DateOptions)

ifDoubleArrayOptions :: Lens' IndexField (Maybe DoubleArrayOptions)

ifDoubleOptions :: Lens' IndexField (Maybe DoubleOptions)

ifIndexFieldName :: Lens' IndexField Text

A string that represents the name of an index field. CloudSearch supports regular index fields as well as dynamic fields. A dynamic field's name defines a pattern that begins or ends with a wildcard. Any document fields that don't map to a regular index field but do match a dynamic field's pattern are configured with the dynamic field's indexing options.

Regular field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Dynamic field names must begin or end with a wildcard (*). The wildcard can also be the only character in a dynamic field name. Multiple wildcards, and wildcards embedded within a string are not supported.

The name score is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

ifIndexFieldType :: Lens' IndexField IndexFieldType

ifIntArrayOptions :: Lens' IndexField (Maybe IntArrayOptions)

ifIntOptions :: Lens' IndexField (Maybe IntOptions)

ifLatLonOptions :: Lens' IndexField (Maybe LatLonOptions)

ifLiteralArrayOptions :: Lens' IndexField (Maybe LiteralArrayOptions)

ifLiteralOptions :: Lens' IndexField (Maybe LiteralOptions)

ifTextArrayOptions :: Lens' IndexField (Maybe TextArrayOptions)

ifTextOptions :: Lens' IndexField (Maybe TextOptions)

DateOptions

data DateOptions

dateOptions :: DateOptions

DateOptions constructor.

The fields accessible through corresponding lenses are:

doDefaultValue :: Maybe Text

doFacetEnabled :: Maybe Bool

doReturnEnabled :: Maybe Bool

doSearchEnabled :: Maybe Bool

doSortEnabled :: Maybe Bool

doSourceField :: Maybe Text

doDefaultValue :: Lens' DateOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

doFacetEnabled :: Lens' DateOptions (Maybe Bool)

Whether facet information can be returned for the field.

doReturnEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

doSearchEnabled :: Lens' DateOptions (Maybe Bool)

Whether the contents of the field are searchable.

doSortEnabled :: Lens' DateOptions (Maybe Bool)

Whether the field can be used to sort the search results.

doSourceField :: Lens' DateOptions (Maybe Text)

OptionState

data OptionState

Constructors

Active
Active

FailedToValidate
FailedToValidate

Processing
Processing

RequiresIndexDocuments
RequiresIndexDocuments

TextArrayOptions

data TextArrayOptions

textArrayOptions :: TextArrayOptions

TextArrayOptions constructor.

The fields accessible through corresponding lenses are:

taoAnalysisScheme :: Maybe Text

taoDefaultValue :: Maybe Text

taoHighlightEnabled :: Maybe Bool

taoReturnEnabled :: Maybe Bool

taoSourceFields :: Maybe Text

taoAnalysisScheme :: Lens' TextArrayOptions (Maybe Text)

The name of an analysis scheme for a 'text-array' field.

taoDefaultValue :: Lens' TextArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

taoHighlightEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether highlights can be returned for the field.

taoReturnEnabled :: Lens' TextArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

taoSourceFields :: Lens' TextArrayOptions (Maybe Text)

A list of source fields to map to the field.

AlgorithmicStemming

data AlgorithmicStemming

Constructors

Full
full

Light
light

Minimal
minimal

None
none

AnalysisScheme

data AnalysisScheme

analysisScheme

Arguments

:: Text
asAnalysisSchemeName

-> AnalysisSchemeLanguage
asAnalysisSchemeLanguage

-> AnalysisScheme

AnalysisScheme constructor.

The fields accessible through corresponding lenses are:

asAnalysisOptions :: Maybe AnalysisOptions

asAnalysisSchemeLanguage :: AnalysisSchemeLanguage

asAnalysisSchemeName :: Text

asAnalysisOptions :: Lens' AnalysisScheme (Maybe AnalysisOptions)

asAnalysisSchemeLanguage :: Lens' AnalysisScheme AnalysisSchemeLanguage

asAnalysisSchemeName :: Lens' AnalysisScheme Text

ScalingParameters

data ScalingParameters

scalingParameters :: ScalingParameters

ScalingParameters constructor.

The fields accessible through corresponding lenses are:

spDesiredInstanceType :: Maybe PartitionInstanceType

spDesiredPartitionCount :: Maybe Natural

spDesiredReplicationCount :: Maybe Natural

spDesiredInstanceType :: Lens' ScalingParameters (Maybe PartitionInstanceType)

The instance type that you want to preconfigure for your domain. For example, 'search.m1.small'.

spDesiredPartitionCount :: Lens' ScalingParameters (Maybe Natural)

The number of partitions you want to preconfigure for your domain. Only valid when you select 'm2.2xlarge' as the desired instance type.

spDesiredReplicationCount :: Lens' ScalingParameters (Maybe Natural)

The number of replicas you want to preconfigure for each index partition.

AnalysisOptions

data AnalysisOptions

analysisOptions :: AnalysisOptions

AnalysisOptions constructor.

The fields accessible through corresponding lenses are:

aoAlgorithmicStemming :: Maybe AlgorithmicStemming

aoJapaneseTokenizationDictionary :: Maybe Text

aoStemmingDictionary :: Maybe Text

aoStopwords :: Maybe Text

aoSynonyms :: Maybe Text

aoAlgorithmicStemming :: Lens' AnalysisOptions (Maybe AlgorithmicStemming)

The level of algorithmic stemming to perform: none, minimal, light, or full. The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearchDeveloper Guide

aoJapaneseTokenizationDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON array that contains a collection of terms, tokens, readings and part of speech for Japanese Tokenizaiton. The Japanese tokenization dictionary enables you to override the default tokenization for selected terms. This is only valid for Japanese language fields.

aoStemmingDictionary :: Lens' AnalysisOptions (Maybe Text)

A JSON object that contains a collection of string:value pairs that each map a term to its stem. For example, '{"term1": "stem1", "term2": "stem2","term3": "stem3"}'. The stemming dictionary is applied in addition to any algorithmic stemming. This enables you to override the results of the algorithmic stemming to correct specific cases of overstemming or understemming. The maximum size of a stemming dictionary is 500 KB.

aoStopwords :: Lens' AnalysisOptions (Maybe Text)

A JSON array of terms to ignore during indexing and searching. For example, '["a", "an", "the", "of"]'. The stopwords dictionary must explicitly list each word you want to ignore. Wildcards and regular expressions are not supported.

aoSynonyms :: Lens' AnalysisOptions (Maybe Text)

A JSON object that defines synonym groups and aliases. A synonym group is an array of arrays, where each sub-array is a group of terms where each term in the group is considered a synonym of every other term in the group. The aliases value is an object that contains a collection of string:value pairs where the string specifies a term and the array of values specifies each of the aliases for that term. An alias is considered a synonym of the specified term, but the term is not considered a synonym of the alias. For more information about specifying synonyms, see Synonyms in the Amazon CloudSearchDeveloper Guide.

DoubleOptions

data DoubleOptions

doubleOptions :: DoubleOptions

DoubleOptions constructor.

The fields accessible through corresponding lenses are:

do1DefaultValue :: Maybe Double

do1FacetEnabled :: Maybe Bool

do1ReturnEnabled :: Maybe Bool

do1SearchEnabled :: Maybe Bool

do1SortEnabled :: Maybe Bool

do1SourceField :: Maybe Text

do1DefaultValue :: Lens' DoubleOptions (Maybe Double)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

do1FacetEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether facet information can be returned for the field.

do1ReturnEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

do1SearchEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the contents of the field are searchable.

do1SortEnabled :: Lens' DoubleOptions (Maybe Bool)

Whether the field can be used to sort the search results.

do1SourceField :: Lens' DoubleOptions (Maybe Text)

The name of the source field to map to the field.

TextOptions

data TextOptions

textOptions :: TextOptions

TextOptions constructor.

The fields accessible through corresponding lenses are:

toAnalysisScheme :: Maybe Text

toDefaultValue :: Maybe Text

toHighlightEnabled :: Maybe Bool

toReturnEnabled :: Maybe Bool

toSortEnabled :: Maybe Bool

toSourceField :: Maybe Text

toAnalysisScheme :: Lens' TextOptions (Maybe Text)

The name of an analysis scheme for a text field.

toDefaultValue :: Lens' TextOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

toHighlightEnabled :: Lens' TextOptions (Maybe Bool)

Whether highlights can be returned for the field.

toReturnEnabled :: Lens' TextOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

toSortEnabled :: Lens' TextOptions (Maybe Bool)

Whether the field can be used to sort the search results.

toSourceField :: Lens' TextOptions (Maybe Text)

AvailabilityOptionsStatus

data AvailabilityOptionsStatus

availabilityOptionsStatus

Arguments

:: Bool
aosOptions

-> OptionStatus
aosStatus

-> AvailabilityOptionsStatus

AvailabilityOptionsStatus constructor.

The fields accessible through corresponding lenses are:

aosOptions :: Bool

aosStatus :: OptionStatus

aosOptions :: Lens' AvailabilityOptionsStatus Bool

The availability options configured for the domain.

aosStatus :: Lens' AvailabilityOptionsStatus OptionStatus

IndexFieldStatus

data IndexFieldStatus

indexFieldStatus

Arguments

:: IndexField
ifsOptions

-> OptionStatus
ifsStatus

-> IndexFieldStatus

IndexFieldStatus constructor.

The fields accessible through corresponding lenses are:

ifsOptions :: IndexField

ifsStatus :: OptionStatus

ifsOptions :: Lens' IndexFieldStatus IndexField

ifsStatus :: Lens' IndexFieldStatus OptionStatus

ScalingParametersStatus

data ScalingParametersStatus

scalingParametersStatus

Arguments

:: ScalingParameters
spsOptions

-> OptionStatus
spsStatus

-> ScalingParametersStatus

ScalingParametersStatus constructor.

The fields accessible through corresponding lenses are:

spsOptions :: ScalingParameters

spsStatus :: OptionStatus

spsOptions :: Lens' ScalingParametersStatus ScalingParameters

spsStatus :: Lens' ScalingParametersStatus OptionStatus

AnalysisSchemeStatus

data AnalysisSchemeStatus

analysisSchemeStatus

Arguments

:: AnalysisScheme
assOptions

-> OptionStatus
assStatus

-> AnalysisSchemeStatus

AnalysisSchemeStatus constructor.

The fields accessible through corresponding lenses are:

assOptions :: AnalysisScheme

assStatus :: OptionStatus

assOptions :: Lens' AnalysisSchemeStatus AnalysisScheme

assStatus :: Lens' AnalysisSchemeStatus OptionStatus

ServiceEndpoint

data ServiceEndpoint

serviceEndpoint :: ServiceEndpoint

ServiceEndpoint constructor.

The fields accessible through corresponding lenses are:

seEndpoint :: Maybe Text

seEndpoint :: Lens' ServiceEndpoint (Maybe Text)

Limits

data Limits

limits

Arguments

:: Natural
lMaximumReplicationCount

-> Natural
lMaximumPartitionCount

-> Limits

Limits constructor.

The fields accessible through corresponding lenses are:

lMaximumPartitionCount :: Natural

lMaximumReplicationCount :: Natural

lMaximumPartitionCount :: Lens' Limits Natural

lMaximumReplicationCount :: Lens' Limits Natural

ExpressionStatus

data ExpressionStatus

expressionStatus

Arguments

:: Expression
esOptions

-> OptionStatus
esStatus

-> ExpressionStatus

ExpressionStatus constructor.

The fields accessible through corresponding lenses are:

esOptions :: Expression

esStatus :: OptionStatus

esOptions :: Lens' ExpressionStatus Expression

The expression that is evaluated for sorting while processing a search request.

esStatus :: Lens' ExpressionStatus OptionStatus

IndexFieldType

data IndexFieldType

Constructors

IFTDate
date

IFTDateArray
date-array

IFTDouble
double

IFTDoubleArray
double-array

IFTInt
int

IFTIntArray
int-array

IFTLatlon
latlon

IFTLiteral
literal

IFTLiteralArray
literal-array

IFTText
text

IFTTextArray
text-array

LatLonOptions

data LatLonOptions

latLonOptions :: LatLonOptions

LatLonOptions constructor.

The fields accessible through corresponding lenses are:

lloDefaultValue :: Maybe Text

lloFacetEnabled :: Maybe Bool

lloReturnEnabled :: Maybe Bool

lloSearchEnabled :: Maybe Bool

lloSortEnabled :: Maybe Bool

lloSourceField :: Maybe Text

lloDefaultValue :: Lens' LatLonOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

lloFacetEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether facet information can be returned for the field.

lloReturnEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

lloSearchEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the contents of the field are searchable.

lloSortEnabled :: Lens' LatLonOptions (Maybe Bool)

Whether the field can be used to sort the search results.

lloSourceField :: Lens' LatLonOptions (Maybe Text)

SuggesterStatus

data SuggesterStatus

suggesterStatus

Arguments

:: Suggester
ssOptions

-> OptionStatus
ssStatus

-> SuggesterStatus

SuggesterStatus constructor.

The fields accessible through corresponding lenses are:

ssOptions :: Suggester

ssStatus :: OptionStatus

ssOptions :: Lens' SuggesterStatus Suggester

ssStatus :: Lens' SuggesterStatus OptionStatus

OptionStatus

data OptionStatus

optionStatus

Arguments

:: UTCTime
osCreationDate

-> UTCTime
osUpdateDate

-> OptionState
osState

-> OptionStatus

OptionStatus constructor.

The fields accessible through corresponding lenses are:

osCreationDate :: UTCTime

osPendingDeletion :: Maybe Bool

osState :: OptionState

osUpdateDate :: UTCTime

osUpdateVersion :: Maybe Natural

osCreationDate :: Lens' OptionStatus UTCTime

A timestamp for when this option was created.

osPendingDeletion :: Lens' OptionStatus (Maybe Bool)

Indicates that the option will be deleted once processing is complete.

osState :: Lens' OptionStatus OptionState

The state of processing a change to an option. Possible values:

RequiresIndexDocuments: the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. Processing: the option's latest value is in the process of being activated. Active: the option's latest value is completely deployed. FailedToValidate: the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

osUpdateDate :: Lens' OptionStatus UTCTime

A timestamp for when this option was last updated.

osUpdateVersion :: Lens' OptionStatus (Maybe Natural)

A unique integer that indicates when this option was last updated.

LiteralArrayOptions

data LiteralArrayOptions

literalArrayOptions :: LiteralArrayOptions

LiteralArrayOptions constructor.

The fields accessible through corresponding lenses are:

laoDefaultValue :: Maybe Text

laoFacetEnabled :: Maybe Bool

laoReturnEnabled :: Maybe Bool

laoSearchEnabled :: Maybe Bool

laoSourceFields :: Maybe Text

laoDefaultValue :: Lens' LiteralArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

laoFacetEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

laoReturnEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

laoSearchEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

laoSourceFields :: Lens' LiteralArrayOptions (Maybe Text)

A list of source fields to map to the field.

IntArrayOptions

data IntArrayOptions

intArrayOptions :: IntArrayOptions

IntArrayOptions constructor.

The fields accessible through corresponding lenses are:

iaoDefaultValue :: Maybe Integer

iaoFacetEnabled :: Maybe Bool

iaoReturnEnabled :: Maybe Bool

iaoSearchEnabled :: Maybe Bool

iaoSourceFields :: Maybe Text

iaoDefaultValue :: Lens' IntArrayOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document.

iaoFacetEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

iaoReturnEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

iaoSearchEnabled :: Lens' IntArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

iaoSourceFields :: Lens' IntArrayOptions (Maybe Text)

A list of source fields to map to the field.

Expression

data Expression

expression

Arguments

:: Text
eExpressionName

-> Text
eExpressionValue

-> Expression

Expression constructor.

The fields accessible through corresponding lenses are:

eExpressionName :: Text

eExpressionValue :: Text

eExpressionName :: Lens' Expression Text

eExpressionValue :: Lens' Expression Text

SuggesterFuzzyMatching

data SuggesterFuzzyMatching

Constructors

SFMHigh
high

SFMLow
low

SFMNone
none

DateArrayOptions

data DateArrayOptions

dateArrayOptions :: DateArrayOptions

DateArrayOptions constructor.

The fields accessible through corresponding lenses are:

dao1DefaultValue :: Maybe Text

dao1FacetEnabled :: Maybe Bool

dao1ReturnEnabled :: Maybe Bool

dao1SearchEnabled :: Maybe Bool

dao1SourceFields :: Maybe Text

dao1DefaultValue :: Lens' DateArrayOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

dao1FacetEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether facet information can be returned for the field.

dao1ReturnEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

dao1SearchEnabled :: Lens' DateArrayOptions (Maybe Bool)

Whether the contents of the field are searchable.

dao1SourceFields :: Lens' DateArrayOptions (Maybe Text)

A list of source fields to map to the field.

AnalysisSchemeLanguage

data AnalysisSchemeLanguage

Constructors

Ar
ar

Bg
bg

Ca
ca

Cs
cs

Da
da

De
de

El
el

En
en

Es
es

Eu
eu

Fa
fa

Fi
fi

Fr
fr

Ga
ga

Gl
gl

He
he

Hi
hi

Hu
hu

Hy
hy

Id
id

It
it

Ja
ja

Ko
ko

Lv
lv

Mul
mul

Nl
nl

No
no

Pt
pt

Ro
ro

Ru
ru

Sv
sv

Th
th

Tr
tr

ZhHans
zh-Hans

ZhHant
zh-Hant

PartitionInstanceType

data PartitionInstanceType

Constructors

SearchM1Large
search.m1.large

SearchM1Small
search.m1.small

SearchM22xlarge
search.m2.2xlarge

SearchM2Xlarge
search.m2.xlarge

SearchM32xlarge
search.m3.2xlarge

SearchM3Large
search.m3.large

SearchM3Medium
search.m3.medium

SearchM3Xlarge
search.m3.xlarge

Suggester

data Suggester

suggester

Arguments

:: Text
sSuggesterName

-> DocumentSuggesterOptions
sDocumentSuggesterOptions

-> Suggester

Suggester constructor.

The fields accessible through corresponding lenses are:

sDocumentSuggesterOptions :: DocumentSuggesterOptions

sSuggesterName :: Text

sDocumentSuggesterOptions :: Lens' Suggester DocumentSuggesterOptions

sSuggesterName :: Lens' Suggester Text

IntOptions

data IntOptions

intOptions :: IntOptions

IntOptions constructor.

The fields accessible through corresponding lenses are:

ioDefaultValue :: Maybe Integer

ioFacetEnabled :: Maybe Bool

ioReturnEnabled :: Maybe Bool

ioSearchEnabled :: Maybe Bool

ioSortEnabled :: Maybe Bool

ioSourceField :: Maybe Text

ioDefaultValue :: Lens' IntOptions (Maybe Integer)

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

ioFacetEnabled :: Lens' IntOptions (Maybe Bool)

Whether facet information can be returned for the field.

ioReturnEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

ioSearchEnabled :: Lens' IntOptions (Maybe Bool)

Whether the contents of the field are searchable.

ioSortEnabled :: Lens' IntOptions (Maybe Bool)

Whether the field can be used to sort the search results.

ioSourceField :: Lens' IntOptions (Maybe Text)

The name of the source field to map to the field.

LiteralOptions

data LiteralOptions

literalOptions :: LiteralOptions

LiteralOptions constructor.

The fields accessible through corresponding lenses are:

loDefaultValue :: Maybe Text

loFacetEnabled :: Maybe Bool

loReturnEnabled :: Maybe Bool

loSearchEnabled :: Maybe Bool

loSortEnabled :: Maybe Bool

loSourceField :: Maybe Text

loDefaultValue :: Lens' LiteralOptions (Maybe Text)

A value to use for the field if the field isn't specified for a document.

loFacetEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether facet information can be returned for the field.

loReturnEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field can be returned in the search results.

loSearchEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the contents of the field are searchable.

loSortEnabled :: Lens' LiteralOptions (Maybe Bool)

Whether the field can be used to sort the search results.

loSourceField :: Lens' LiteralOptions (Maybe Text)

AccessPoliciesStatus

data AccessPoliciesStatus

accessPoliciesStatus

Arguments

:: Text
apsOptions

-> OptionStatus
apsStatus

-> AccessPoliciesStatus

AccessPoliciesStatus constructor.

The fields accessible through corresponding lenses are:

apsOptions :: Text

apsStatus :: OptionStatus

apsOptions :: Lens' AccessPoliciesStatus Text

apsStatus :: Lens' AccessPoliciesStatus OptionStatus