Network.AWS.CloudSearch.UpdateScalingParameters

Description

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide .

Creating a Request

updateScalingParameters

Arguments

:: Text
uspDomainName

-> ScalingParameters
uspScalingParameters

-> UpdateScalingParameters

Creates a value of UpdateScalingParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uspDomainName - Undocumented member.

uspScalingParameters - Undocumented member.

data UpdateScalingParameters

Container for the parameters to the UpdateScalingParameters operation. Specifies the name of the domain you want to update and the scaling parameters you want to configure.

See: updateScalingParameters smart constructor.

Request Lenses

uspDomainName :: Lens' UpdateScalingParameters Text

Undocumented member.

uspScalingParameters :: Lens' UpdateScalingParameters ScalingParameters

Undocumented member.

Destructuring the Response

updateScalingParametersResponse

Arguments

:: Int
usprsResponseStatus

-> ScalingParametersStatus
usprsScalingParameters

-> UpdateScalingParametersResponse

Creates a value of UpdateScalingParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usprsResponseStatus - -- | The response status code.

usprsScalingParameters - Undocumented member.

data UpdateScalingParametersResponse

The result of a UpdateScalingParameters request. Contains the status of the newly-configured scaling parameters.

See: updateScalingParametersResponse smart constructor.

Response Lenses

usprsResponseStatus :: Lens' UpdateScalingParametersResponse Int

- | The response status code.

usprsScalingParameters :: Lens' UpdateScalingParametersResponse ScalingParametersStatus

Undocumented member.