Network.AWS.CloudSearch.UpdateServiceAccessPolicies

Description

Configures the access rules that control access to the domain's document and search endpoints. For more information, see Configuring Access for an Amazon CloudSearch Domain .

Creating a Request

updateServiceAccessPolicies

Arguments

:: Text
usapDomainName

-> Text
usapAccessPolicies

-> UpdateServiceAccessPolicies

Creates a value of UpdateServiceAccessPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usapDomainName - Undocumented member.

usapAccessPolicies - The access rules you want to configure. These rules replace any existing rules.

data UpdateServiceAccessPolicies

Container for the parameters to the UpdateServiceAccessPolicies operation. Specifies the name of the domain you want to update and the access rules you want to configure.

See: updateServiceAccessPolicies smart constructor.

Request Lenses

usapDomainName :: Lens' UpdateServiceAccessPolicies Text

Undocumented member.

usapAccessPolicies :: Lens' UpdateServiceAccessPolicies Text

The access rules you want to configure. These rules replace any existing rules.

Destructuring the Response

updateServiceAccessPoliciesResponse

Arguments

:: Int
usaprsResponseStatus

-> AccessPoliciesStatus
usaprsAccessPolicies

-> UpdateServiceAccessPoliciesResponse

Creates a value of UpdateServiceAccessPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usaprsResponseStatus - -- | The response status code.

usaprsAccessPolicies - The access rules configured for the domain.

data UpdateServiceAccessPoliciesResponse

The result of an UpdateServiceAccessPolicies request. Contains the new access policies.

See: updateServiceAccessPoliciesResponse smart constructor.

Response Lenses

usaprsResponseStatus :: Lens' UpdateServiceAccessPoliciesResponse Int

- | The response status code.

usaprsAccessPolicies :: Lens' UpdateServiceAccessPoliciesResponse AccessPoliciesStatus

The access rules configured for the domain.