Network.AWS.CloudSearch.V2013_01_01

Description

Amazon CloudSearch is a fully-managed service in the AWS Cloud that makes it simple and cost-effective to set up, manage, and scale a search solution for your website or application. Amazon CloudSearch supports 34 languages and popular search features such as highlighting, autocomplete, and geospatial search.

Documentation

data BuildSuggesters

Constructors

BuildSuggesters

Fields

_bsrDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

bsrDomainName :: Iso' BuildSuggesters Text

data BuildSuggestersResponse

Constructors

BuildSuggestersResponse

Fields

_bssFieldNames :: [Text]

A list of field names.

bssFieldNames :: Iso' BuildSuggestersResponse [Text]

data CreateDomain

Constructors

CreateDomain

Fields

_cdrDomainName :: Text

A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

cdrDomainName :: Iso' CreateDomain Text

data CreateDomainResponse

Constructors

CreateDomainResponse

Fields

_cdsDomainStatus :: Maybe DomainStatus

The current status of the search domain.

cdsDomainStatus :: Iso' CreateDomainResponse (Maybe DomainStatus)

data DefineAnalysisScheme

Constructors

DefineAnalysisScheme

Fields

_dasvAnalysisScheme :: AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

_dasvDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dasvDomainName :: Lens' DefineAnalysisScheme Text

dasvAnalysisScheme :: Lens' DefineAnalysisScheme AnalysisScheme

data DefineAnalysisSchemeResponse

Constructors

DefineAnalysisSchemeResponse

Fields

_daswAnalysisScheme :: AnalysisSchemeStatus

The status and configuration of an AnalysisScheme.

daswAnalysisScheme :: Iso' DefineAnalysisSchemeResponse AnalysisSchemeStatus

data DefineExpression

Constructors

DefineExpression

Fields

_detDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_detExpression :: Expression

A named expression that can be evaluated at search time. Can be used for sorting and filtering search results and constructing other expressions.

detExpression :: Lens' DefineExpression Expression

detDomainName :: Lens' DefineExpression Text

data DefineExpressionResponse

Constructors

DefineExpressionResponse

Fields

_deuExpression :: ExpressionStatus

The value of an Expression and its current status.

deuExpression :: Iso' DefineExpressionResponse ExpressionStatus

data DefineIndexField

Constructors

DefineIndexField

Fields

_difvDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_difvIndexField :: IndexField

The index field and field options you want to configure.

difvIndexField :: Lens' DefineIndexField IndexField

difvDomainName :: Lens' DefineIndexField Text

data DefineIndexFieldResponse

Constructors

DefineIndexFieldResponse

Fields

_difwIndexField :: IndexFieldStatus

The value of an IndexField and its current status.

difwIndexField :: Iso' DefineIndexFieldResponse IndexFieldStatus

data DefineSuggester

Constructors

DefineSuggester

Fields

_dstDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_dstSuggester :: Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

dstSuggester :: Lens' DefineSuggester Suggester

dstDomainName :: Lens' DefineSuggester Text

data DefineSuggesterResponse

Constructors

DefineSuggesterResponse

Fields

_dsuSuggester :: SuggesterStatus

The value of a Suggester and its current status.

dsuSuggester :: Iso' DefineSuggesterResponse SuggesterStatus

data DeleteAnalysisScheme

Constructors

DeleteAnalysisScheme

Fields

_dasrDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_dasrAnalysisSchemeName :: Text

The name of the analysis scheme you want to delete.

dasrDomainName :: Lens' DeleteAnalysisScheme Text

dasrAnalysisSchemeName :: Lens' DeleteAnalysisScheme Text

data DeleteAnalysisSchemeResponse

Constructors

DeleteAnalysisSchemeResponse

Fields

_dassAnalysisScheme :: AnalysisSchemeStatus

The status of the analysis scheme being deleted.

dassAnalysisScheme :: Iso' DeleteAnalysisSchemeResponse AnalysisSchemeStatus

data DeleteDomain

Constructors

DeleteDomain

Fields

_ddyDomainName :: Text

The name of the domain you want to permanently delete.

ddyDomainName :: Iso' DeleteDomain Text

data DeleteDomainResponse

Constructors

DeleteDomainResponse

Fields

_ddzDomainStatus :: Maybe DomainStatus

The current status of the search domain.

ddzDomainStatus :: Iso' DeleteDomainResponse (Maybe DomainStatus)

data DeleteExpression

Constructors

DeleteExpression

Fields

_devDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_devExpressionName :: Text

The name of the Expression to delete.

devExpressionName :: Lens' DeleteExpression Text

devDomainName :: Lens' DeleteExpression Text

data DeleteExpressionResponse

Constructors

DeleteExpressionResponse

Fields

_dewExpression :: ExpressionStatus

The status of the expression being deleted.

dewExpression :: Iso' DeleteExpressionResponse ExpressionStatus

data DeleteIndexField

Constructors

DeleteIndexField

Fields

_diftDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_diftIndexFieldName :: Text

The name of the index field your want to remove from the domain's indexing options.

diftIndexFieldName :: Lens' DeleteIndexField Text

diftDomainName :: Lens' DeleteIndexField Text

data DeleteIndexFieldResponse

Constructors

DeleteIndexFieldResponse

Fields

_difuIndexField :: IndexFieldStatus

The status of the index field being deleted.

difuIndexField :: Iso' DeleteIndexFieldResponse IndexFieldStatus

data DeleteSuggester

Constructors

DeleteSuggester

Fields

_dsvDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_dsvSuggesterName :: Text

Specifies the name of the suggester you want to delete.

dsvSuggesterName :: Lens' DeleteSuggester Text

dsvDomainName :: Lens' DeleteSuggester Text

data DeleteSuggesterResponse

Constructors

DeleteSuggesterResponse

Fields

_dswSuggester :: SuggesterStatus

The status of the suggester being deleted.

dswSuggester :: Iso' DeleteSuggesterResponse SuggesterStatus

describeAnalysisSchemes

Arguments

:: Text
_dastDomainName

-> DescribeAnalysisSchemes

Minimum specification for a DescribeAnalysisSchemes request.

data DescribeAnalysisSchemes

Constructors

DescribeAnalysisSchemes

Fields

_dastDomainName :: Text

The name of the domain you want to describe.

_dastDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_dastAnalysisSchemeNames :: [Text]

The analysis schemes you want to describe.

dastDomainName :: Lens' DescribeAnalysisSchemes Text

dastDeployed :: Lens' DescribeAnalysisSchemes (Maybe Bool)

dastAnalysisSchemeNames :: Lens' DescribeAnalysisSchemes [Text]

data DescribeAnalysisSchemesResponse

Constructors

DescribeAnalysisSchemesResponse

Fields

_dasuAnalysisSchemes :: [AnalysisSchemeStatus]

The analysis scheme descriptions.

dasuAnalysisSchemes :: Iso' DescribeAnalysisSchemesResponse [AnalysisSchemeStatus]

describeAvailabilityOptions

Arguments

:: Text
_daorDomainName

-> DescribeAvailabilityOptions

Minimum specification for a DescribeAvailabilityOptions request.

data DescribeAvailabilityOptions

Constructors

DescribeAvailabilityOptions

Fields

_daorDomainName :: Text

The name of the domain you want to describe.

_daorDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

daorDomainName :: Lens' DescribeAvailabilityOptions Text

daorDeployed :: Lens' DescribeAvailabilityOptions (Maybe Bool)

data DescribeAvailabilityOptionsResponse

Constructors

DescribeAvailabilityOptionsResponse

Fields

_daosAvailabilityOptions :: Maybe AvailabilityOptionsStatus

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

daosAvailabilityOptions :: Iso' DescribeAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

describeDomains :: DescribeDomains

Minimum specification for a DescribeDomains request.

data DescribeDomains

Constructors

DescribeDomains

Fields

_ddrDomainNames :: [Text]

The names of the domains you want to include in the response.

ddrDomainNames :: Iso' DescribeDomains [Text]

data DescribeDomainsResponse

Constructors

DescribeDomainsResponse

Fields

_ddsDomainStatusList :: [DomainStatus]

A list that contains the status of each requested domain.

ddsDomainStatusList :: Iso' DescribeDomainsResponse [DomainStatus]

describeExpressions

Arguments

:: Text
_derDomainName

-> DescribeExpressions

Minimum specification for a DescribeExpressions request.

data DescribeExpressions

Constructors

DescribeExpressions

Fields

_derDomainName :: Text

The name of the domain you want to describe.

_derDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_derExpressionNames :: [Text]

Limits the DescribeExpressions response to the specified expressions. If not specified, all expressions are shown.

derExpressionNames :: Lens' DescribeExpressions [Text]

derDomainName :: Lens' DescribeExpressions Text

derDeployed :: Lens' DescribeExpressions (Maybe Bool)

data DescribeExpressionsResponse

Constructors

DescribeExpressionsResponse

Fields

_desExpressions :: [ExpressionStatus]

The expressions configured for the domain.

desExpressions :: Iso' DescribeExpressionsResponse [ExpressionStatus]

describeIndexFields

Arguments

:: Text
_difrDomainName

-> DescribeIndexFields

Minimum specification for a DescribeIndexFields request.

data DescribeIndexFields

Constructors

DescribeIndexFields

Fields

_difrDomainName :: Text

The name of the domain you want to describe.

_difrDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_difrFieldNames :: [Text]

A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

difrFieldNames :: Lens' DescribeIndexFields [Text]

difrDomainName :: Lens' DescribeIndexFields Text

difrDeployed :: Lens' DescribeIndexFields (Maybe Bool)

data DescribeIndexFieldsResponse

Constructors

DescribeIndexFieldsResponse

Fields

_difsIndexFields :: [IndexFieldStatus]

The index fields configured for the domain.

difsIndexFields :: Iso' DescribeIndexFieldsResponse [IndexFieldStatus]

data DescribeScalingParameters

Constructors

DescribeScalingParameters

Fields

_dsprDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

dsprDomainName :: Iso' DescribeScalingParameters Text

data DescribeScalingParametersResponse

Constructors

DescribeScalingParametersResponse

Fields

_dspsScalingParameters :: ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

dspsScalingParameters :: Iso' DescribeScalingParametersResponse ScalingParametersStatus

describeServiceAccessPolicies

Arguments

:: Text
_dsaprDomainName

-> DescribeServiceAccessPolicies

Minimum specification for a DescribeServiceAccessPolicies request.

data DescribeServiceAccessPolicies

Constructors

DescribeServiceAccessPolicies

Fields

_dsaprDomainName :: Text

The name of the domain you want to describe.

_dsaprDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

dsaprDomainName :: Lens' DescribeServiceAccessPolicies Text

dsaprDeployed :: Lens' DescribeServiceAccessPolicies (Maybe Bool)

data DescribeServiceAccessPoliciesResponse

Constructors

DescribeServiceAccessPoliciesResponse

Fields

_dsapsAccessPolicies :: AccessPoliciesStatus

The access rules configured for the domain specified in the request.

dsapsAccessPolicies :: Iso' DescribeServiceAccessPoliciesResponse AccessPoliciesStatus

describeSuggesters

Arguments

:: Text
_dsrDomainName

-> DescribeSuggesters

Minimum specification for a DescribeSuggesters request.

data DescribeSuggesters

Constructors

DescribeSuggesters

Fields

_dsrDomainName :: Text

The name of the domain you want to describe.

_dsrDeployed :: Maybe Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

_dsrSuggesterNames :: [Text]

The suggesters you want to describe.

dsrSuggesterNames :: Lens' DescribeSuggesters [Text]

dsrDomainName :: Lens' DescribeSuggesters Text

dsrDeployed :: Lens' DescribeSuggesters (Maybe Bool)

data DescribeSuggestersResponse

Constructors

DescribeSuggestersResponse

Fields

_dssSuggesters :: [SuggesterStatus]

The suggesters configured for the domain specified in the request.

dssSuggesters :: Iso' DescribeSuggestersResponse [SuggesterStatus]

data IndexDocuments

Constructors

IndexDocuments

Fields

_idrDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

idrDomainName :: Iso' IndexDocuments Text

data IndexDocumentsResponse

Constructors

IndexDocumentsResponse

Fields

_idsFieldNames :: [Text]

The names of the fields that are currently being indexed.

idsFieldNames :: Iso' IndexDocumentsResponse [Text]

data ListDomainNames

Constructors

ListDomainNames

data ListDomainNamesResponse

Constructors

ListDomainNamesResponse

Fields

_ldnrDomainNames :: Map Text Text

The names of the search domains owned by an account.

ldnrDomainNames :: Iso' ListDomainNamesResponse (Map Text Text)

data CloudSearch

Supported version (2013-01-01) of the Amazon CloudSearch service.

xmlOptions :: Tagged a XMLOptions

data AlgorithmicStemming

The level of algorithmic stemming to perform: none, minimal, light, or full. The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearch Developer Guide.

Constructors

AlgorithmicStemmingFull
full

AlgorithmicStemmingLight
light

AlgorithmicStemmingMinimal
minimal

AlgorithmicStemmingNone
none

data AnalysisSchemeLanguage

An IETF RFC 4646 language code or mul for multiple languages.

Constructors

AnalysisSchemeLanguageAr
ar

AnalysisSchemeLanguageBg
bg

AnalysisSchemeLanguageCa
ca

AnalysisSchemeLanguageCs
cs

AnalysisSchemeLanguageDa
da

AnalysisSchemeLanguageDe
de

AnalysisSchemeLanguageEl
el

AnalysisSchemeLanguageEn
en

AnalysisSchemeLanguageEs
es

AnalysisSchemeLanguageEu
eu

AnalysisSchemeLanguageFa
fa

AnalysisSchemeLanguageFi
fi

AnalysisSchemeLanguageFr
fr

AnalysisSchemeLanguageGa
ga

AnalysisSchemeLanguageGl
gl

AnalysisSchemeLanguageHe
he

AnalysisSchemeLanguageHi
hi

AnalysisSchemeLanguageHu
hu

AnalysisSchemeLanguageHy
hy

AnalysisSchemeLanguageId
id

AnalysisSchemeLanguageIt
it

AnalysisSchemeLanguageJa
ja

AnalysisSchemeLanguageKo
ko

AnalysisSchemeLanguageLv
lv

AnalysisSchemeLanguageMul
mul

AnalysisSchemeLanguageNl
nl

AnalysisSchemeLanguageNo
no

AnalysisSchemeLanguagePt
pt

AnalysisSchemeLanguageRo
ro

AnalysisSchemeLanguageRu
ru

AnalysisSchemeLanguageSv
sv

AnalysisSchemeLanguageTh
th

AnalysisSchemeLanguageTr
tr

AnalysisSchemeLanguageZhHans
zh-Hans

AnalysisSchemeLanguageZhHant
zh-Hant

data IndexFieldType

The type of field. The valid options for a field depend on the field type. For more information about the supported field types, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Constructors

IndexFieldTypeDate
date

IndexFieldTypeDateArray
date-array

IndexFieldTypeDouble
double

IndexFieldTypeDoubleArray
double-array

IndexFieldTypeInt
int

IndexFieldTypeIntArray
int-array

IndexFieldTypeLatlon
latlon

IndexFieldTypeLiteral
literal

IndexFieldTypeLiteralArray
literal-array

IndexFieldTypeText
text

IndexFieldTypeTextArray
text-array

data OptionState

The state of processing a change to an option. Possible values: RequiresIndexDocuments: the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. Processing: the option's latest value is in the process of being activated. Active: the option's latest value is completely deployed. FailedToValidate: the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

Constructors

OptionStateActive
Active

OptionStateFailedToValidate
FailedToValidate

OptionStateProcessing
Processing

OptionStateRequiresIndexDocuments
RequiresIndexDocuments

data PartitionInstanceType

The instance type that you want to preconfigure for your domain. For example, search.m1.small.

Constructors

PartitionInstanceTypeSearchM1Large
search.m1.large

PartitionInstanceTypeSearchM1Small
search.m1.small

PartitionInstanceTypeSearchM22Xlarge
search.m2.2xlarge

PartitionInstanceTypeSearchM2Xlarge
search.m2.xlarge

data SuggesterFuzzyMatching

The level of fuzziness allowed when suggesting matches for a string: none, low, or high. With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

Constructors

SuggesterFuzzyMatchingHigh
high

SuggesterFuzzyMatchingLow
low

SuggesterFuzzyMatchingNone
none

newtype ServiceEndpoint

The service endpoint for updating documents in a search domain.

Constructors

ServiceEndpoint

Fields

_seEndpoint :: Maybe Text

The endpoint to which service requests can be submitted. For example, search-imdb-movies-oopcnjfn6ugofer3zx5iadxxca.eu-west-1.cloudsearch.amazonaws.com or doc-imdb-movies-oopcnjfn6ugofer3zx5iadxxca.eu-west-1.cloudsearch.amazonaws.com.

data AccessPoliciesStatus

The access rules configured for the domain specified in the request.

Constructors

AccessPoliciesStatus

Fields

_apsStatus :: OptionStatus

The status of domain configuration option.

_apsOptions :: Text

Access rules for a domain's document or search service endpoints. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide. The maximum size of a policy document is 100 KB.

data AnalysisOptions

Synonyms, stopwords, and stemming options for an analysis scheme.

Constructors

AnalysisOptions

Fields

_aoAlgorithmicStemming :: Maybe AlgorithmicStemming

The level of algorithmic stemming to perform: none, minimal, light, or full. The available levels vary depending on the language. For more information, see Language Specific Text Processing Settings in the Amazon CloudSearch Developer Guide.

_aoStopwords :: Maybe Text

A JSON array of terms to ignore during indexing and searching. For example, ["a", "an", "the", "of"]. The stopwords dictionary must explicitly list each word you want to ignore. Wildcards and regular expressions are not supported.

_aoSynonyms :: Maybe Text

A JSON object that defines synonym groups and aliases. A synonym group is an array of arrays, where each sub-array is a group of terms where each term in the group is considered a synonym of every other term in the group. The aliases value is an object that contains a collection of string:value pairs where the string specifies a term and the array of values specifies each of the aliases for that term. An alias is considered a synonym of the specified term, but the term is not considered a synonym of the alias. For more information about specifying synonyms, see Synonyms in the Amazon CloudSearch Developer Guide.

_aoStemmingDictionary :: Maybe Text

A JSON object that contains a collection of string:value pairs that each map a term to its stem. For example, {"term1": "stem1", "term2": "stem2", "term3": "stem3"}. The stemming dictionary is applied in addition to any algorithmic stemming. This enables you to override the results of the algorithmic stemming to correct specific cases of overstemming or understemming. The maximum size of a stemming dictionary is 500 KB.

data AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

Constructors

AnalysisScheme

Fields

_asAnalysisOptions :: Maybe AnalysisOptions

Synonyms, stopwords, and stemming options for an analysis scheme.

_asAnalysisSchemeLanguage :: AnalysisSchemeLanguage

An IETF RFC 4646 language code or mul for multiple languages.

_asAnalysisSchemeName :: Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

data AnalysisSchemeStatus

The status of the analysis scheme being deleted.

Constructors

AnalysisSchemeStatus

Fields

_assStatus :: OptionStatus

The status of domain configuration option.

_assOptions :: AnalysisScheme

Configuration information for an analysis scheme. Each analysis scheme has a unique name and specifies the language of the text to be processed. The following options can be configured for an analysis scheme: Synonyms, Stopwords, StemmingDictionary, and AlgorithmicStemming.

data AvailabilityOptionsStatus

The availability options configured for the domain. Indicates whether Multi-AZ is enabled for the domain.

Constructors

AvailabilityOptionsStatus

Fields

_aosStatus :: OptionStatus

The status of domain configuration option.

_aosOptions :: Bool

The availability options configured for the domain.

data DateArrayOptions

Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array. All options are enabled by default.

Constructors

DateArrayOptions

Fields

_dapSourceFields :: Maybe Text

A list of source fields to map to the field.

_dapReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_dapFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_dapSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_dapDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

data DateOptions

Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date. All options are enabled by default.

Constructors

DateOptions

Fields

_dduSourceField :: Maybe Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

_dduReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_dduFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_dduSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_dduSortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

_dduDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

data DocumentSuggesterOptions

Options for a search suggester.

Constructors

DocumentSuggesterOptions

Fields

_dsoSourceField :: Text

The name of the index field you want to use for suggestions.

_dsoSortExpression :: Maybe Text

An expression that computes a score for each suggestion to control how they are sorted.

_dsoFuzzyMatching :: Maybe SuggesterFuzzyMatching

The level of fuzziness allowed when suggesting matches for a string: none, low, or high. With none, the specified string is treated as an exact prefix. With low, suggestions must differ from the specified string by no more than one character. With high, suggestions can differ by up to two characters. The default is none.

data DomainStatus

The current status of the search domain.

Constructors

DomainStatus

Fields

_dySearchInstanceCount :: Maybe Integer

The number of search instances that are available to process search requests.

_dySearchInstanceType :: Maybe Text

The instance type that is being used to process search requests.

_dyDocService :: Maybe ServiceEndpoint

The service endpoint for updating documents in a search domain.

_dyARN :: Maybe Text

The Amazon Resource Name (ARN) of the search domain. See Identifiers for IAM Entities in Using AWS Identity and Access Management for more information.

_dyCreated :: Maybe Bool

True if the search domain is created. It can take several minutes to initialize a domain when CreateDomain is called. Newly created search domains are returned from DescribeDomains with a false value for Created until domain creation is complete.

_dySearchService :: Maybe ServiceEndpoint

The service endpoint for requesting search results from a search domain.

_dyRequiresIndexDocuments :: Bool

True if IndexDocuments needs to be called to activate the current domain configuration.

_dyDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_dySearchPartitionCount :: Maybe Integer

The number of partitions across which the search index is spread.

_dyDeleted :: Maybe Bool

True if the search domain has been deleted. The system must clean up resources dedicated to the search domain when DeleteDomain is called. Newly deleted search domains are returned from DescribeDomains with a true value for IsDeleted for several minutes until resource cleanup is complete.

_dyDomainId :: Text

An internally generated unique identifier for a domain.

_dyProcessing :: Maybe Bool

True if processing is being done to activate the current domain configuration.

data DoubleArrayOptions

Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array. All options are enabled by default.

Constructors

DoubleArrayOptions

Fields

_daoSourceFields :: Maybe Text

A list of source fields to map to the field.

_daoReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_daoFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_daoSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_daoDefaultValue :: Maybe Double

A value to use for the field if the field isn't specified for a document.

data DoubleOptions

Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double. All options are enabled by default.

Constructors

DoubleOptions

Fields

_ddvSourceField :: Maybe Text

The name of the source field to map to the field.

_ddvReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_ddvFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_ddvSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_ddvSortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

_ddvDefaultValue :: Maybe Double

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

data Expression

The expression that is evaluated for sorting or filtering while processing a search request.

Constructors

Expression

Fields

_fExpressionName :: Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

_fExpressionValue :: Text

The expression to evaluate for sorting while processing a search request. The Expression syntax is based on JavaScript expressions. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

data ExpressionStatus

The value of an Expression and its current status.

Constructors

ExpressionStatus

Fields

_esStatus :: OptionStatus

The status of domain configuration option.

_esOptions :: Expression

The expression that is evaluated for sorting or filtering while processing a search request.

data IndexField

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

Constructors

IndexField

Fields

_ifDoubleArrayOptions :: Maybe DoubleArrayOptions

Options for a field that contains an array of double-precision 64-bit floating point values. Present if IndexFieldType specifies the field is of type double-array. All options are enabled by default.

_ifDateOptions :: Maybe DateOptions

Options for a date field. Dates and times are specified in UTC (Coordinated Universal Time) according to IETF RFC3339: yyyy-mm-ddT00:00:00Z. Present if IndexFieldType specifies the field is of type date. All options are enabled by default.

_ifTextArrayOptions :: Maybe TextArrayOptions

Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array. A text-array field is always searchable. All options are enabled by default.

_ifDoubleOptions :: Maybe DoubleOptions

Options for a double-precision 64-bit floating point field. Present if IndexFieldType specifies the field is of type double. All options are enabled by default.

_ifTextOptions :: Maybe TextOptions

Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

_ifIndexFieldType :: IndexFieldType

The type of field. The valid options for a field depend on the field type. For more information about the supported field types, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

_ifLatLonOptions :: Maybe LatLonOptions

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

_ifLiteralArrayOptions :: Maybe LiteralArrayOptions

Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array. All options are enabled by default.

_ifIntArrayOptions :: Maybe IntArrayOptions

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

_ifDateArrayOptions :: Maybe DateArrayOptions

Options for a field that contains an array of dates. Present if IndexFieldType specifies the field is of type date-array. All options are enabled by default.

_ifIntOptions :: Maybe IntOptions

Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int. All options are enabled by default.

_ifLiteralOptions :: Maybe LiteralOptions

Options for literal field. Present if IndexFieldType specifies the field is of type literal. All options are enabled by default.

_ifIndexFieldName :: Text

The name of a field in the search index. Field names must begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). Uppercase letters and hyphens are not allowed. The name "score" is reserved and cannot be specified as field or expression name.

data IndexFieldStatus

The value of an IndexField and its current status.

Constructors

IndexFieldStatus

Fields

_ifsStatus :: OptionStatus

The status of domain configuration option.

_ifsOptions :: IndexField

Configuration information for a field in the index, including its name, type, and options. The supported options depend on the IndexFieldType.

data IntArrayOptions

Options for a field that contains an array of 64-bit signed integers. Present if IndexFieldType specifies the field is of type int-array. All options are enabled by default.

Constructors

IntArrayOptions

Fields

_iaoSourceFields :: Maybe Text

A list of source fields to map to the field.

_iaoReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_iaoFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_iaoSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_iaoDefaultValue :: Maybe Integer

A value to use for the field if the field isn't specified for a document.

data IntOptions

Options for a 64-bit signed integer field. Present if IndexFieldType specifies the field is of type int. All options are enabled by default.

Constructors

IntOptions

Fields

_ioSourceField :: Maybe Text

The name of the source field to map to the field.

_ioReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_ioFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_ioSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_ioSortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

_ioDefaultValue :: Maybe Integer

A value to use for the field if the field isn't specified for a document. This can be important if you are using the field in an expression and that field is not present in every document.

data LatLonOptions

Options for a latlon field. A latlon field contains a location stored as a latitude and longitude value pair. Present if IndexFieldType specifies the field is of type latlon. All options are enabled by default.

Constructors

LatLonOptions

Fields

_lloSourceField :: Maybe Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

_lloReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_lloFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_lloSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_lloSortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

_lloDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

data LiteralArrayOptions

Options for a field that contains an array of literal strings. Present if IndexFieldType specifies the field is of type literal-array. All options are enabled by default.

Constructors

LiteralArrayOptions

Fields

_laoSourceFields :: Maybe Text

A list of source fields to map to the field.

_laoReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_laoFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_laoSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_laoDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

data LiteralOptions

Options for literal field. Present if IndexFieldType specifies the field is of type literal. All options are enabled by default.

Constructors

LiteralOptions

Fields

_loSourceField :: Maybe Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

_loReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_loFacetEnabled :: Maybe Bool

Whether facet information can be returned for the field.

_loSearchEnabled :: Maybe Bool

Whether the contents of the field are searchable.

_loSortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

_loDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

data OptionStatus

The status of domain configuration option.

Constructors

OptionStatus

Fields

_osState :: OptionState

The state of processing a change to an option. Possible values: RequiresIndexDocuments: the option's latest value will not be deployed until IndexDocuments has been called and indexing is complete. Processing: the option's latest value is in the process of being activated. Active: the option's latest value is completely deployed. FailedToValidate: the option value is not compatible with the domain's data and cannot be used to index the data. You must either modify the option value or update or remove the incompatible documents.

_osUpdateDate :: ISO8601

A timestamp for when this option was last updated.

_osPendingDeletion :: Maybe Bool

Indicates that the option will be deleted once processing is complete.

_osCreationDate :: ISO8601

A timestamp for when this option was created.

_osUpdateVersion :: Maybe Integer

A unique integer that indicates when this option was last updated.

data ScalingParameters

The desired instance type and desired number of replicas of each index partition.

Constructors

ScalingParameters

Fields

_spDesiredInstanceType :: Maybe PartitionInstanceType

The instance type that you want to preconfigure for your domain. For example, search.m1.small.

_spDesiredReplicationCount :: Maybe Integer

The number of replicas you want to preconfigure for each index partition.

_spDesiredPartitionCount :: Maybe Integer

The number of partitions you want to preconfigure for your domain. Only valid when you select m2.2xlarge as the desired instance type.

data ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

Constructors

ScalingParametersStatus

Fields

_spsStatus :: OptionStatus

The status of domain configuration option.

_spsOptions :: ScalingParameters

The desired instance type and desired number of replicas of each index partition.

data Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

Constructors

Suggester

Fields

_sDocumentSuggesterOptions :: DocumentSuggesterOptions

Options for a search suggester.

_sSuggesterName :: Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

data SuggesterStatus

The value of a Suggester and its current status.

Constructors

SuggesterStatus

Fields

_ssStatus :: OptionStatus

The status of domain configuration option.

_ssOptions :: Suggester

Configuration information for a search suggester. Each suggester has a unique name and specifies the text field you want to use for suggestions. The following options can be configured for a suggester: FuzzyMatching, SortExpression.

data TextArrayOptions

Options for a field that contains an array of text strings. Present if IndexFieldType specifies the field is of type text-array. A text-array field is always searchable. All options are enabled by default.

Constructors

TextArrayOptions

Fields

_taoSourceFields :: Maybe Text

A list of source fields to map to the field.

_taoReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_taoAnalysisScheme :: Maybe Text

The name of an analysis scheme for a text-array field.

_taoHighlightEnabled :: Maybe Bool

Whether highlights can be returned for the field.

_taoDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

data TextOptions

Options for text field. Present if IndexFieldType specifies the field is of type text. A text field is always searchable. All options are enabled by default.

Constructors

TextOptions

Fields

_toSourceField :: Maybe Text

A string that represents the name of an index field. Field names begin with a letter and can contain the following characters: a-z (lowercase), 0-9, and _ (underscore). The name "score" is reserved and cannot be used as a field name. To reference a document's ID, you can use the name _id.

_toReturnEnabled :: Maybe Bool

Whether the contents of the field can be returned in the search results.

_toAnalysisScheme :: Maybe Text

The name of an analysis scheme for a text field.

_toHighlightEnabled :: Maybe Bool

Whether highlights can be returned for the field.

_toSortEnabled :: Maybe Bool

Whether the field can be used to sort the search results.

_toDefaultValue :: Maybe Text

A value to use for the field if the field isn't specified for a document.

seEndpoint :: Iso' ServiceEndpoint (Maybe Text)

apsStatus :: Lens' AccessPoliciesStatus OptionStatus

apsOptions :: Lens' AccessPoliciesStatus Text

aoSynonyms :: Lens' AnalysisOptions (Maybe Text)

aoStopwords :: Lens' AnalysisOptions (Maybe Text)

aoStemmingDictionary :: Lens' AnalysisOptions (Maybe Text)

aoAlgorithmicStemming :: Lens' AnalysisOptions (Maybe AlgorithmicStemming)

asAnalysisSchemeName :: Lens' AnalysisScheme Text

asAnalysisSchemeLanguage :: Lens' AnalysisScheme AnalysisSchemeLanguage

asAnalysisOptions :: Lens' AnalysisScheme (Maybe AnalysisOptions)

assStatus :: Lens' AnalysisSchemeStatus OptionStatus

assOptions :: Lens' AnalysisSchemeStatus AnalysisScheme

aosStatus :: Lens' AvailabilityOptionsStatus OptionStatus

aosOptions :: Lens' AvailabilityOptionsStatus Bool

dapSourceFields :: Lens' DateArrayOptions (Maybe Text)

dapSearchEnabled :: Lens' DateArrayOptions (Maybe Bool)

dapReturnEnabled :: Lens' DateArrayOptions (Maybe Bool)

dapFacetEnabled :: Lens' DateArrayOptions (Maybe Bool)

dapDefaultValue :: Lens' DateArrayOptions (Maybe Text)

dduSourceField :: Lens' DateOptions (Maybe Text)

dduSortEnabled :: Lens' DateOptions (Maybe Bool)

dduSearchEnabled :: Lens' DateOptions (Maybe Bool)

dduReturnEnabled :: Lens' DateOptions (Maybe Bool)

dduFacetEnabled :: Lens' DateOptions (Maybe Bool)

dduDefaultValue :: Lens' DateOptions (Maybe Text)

dsoSourceField :: Lens' DocumentSuggesterOptions Text

dsoSortExpression :: Lens' DocumentSuggesterOptions (Maybe Text)

dsoFuzzyMatching :: Lens' DocumentSuggesterOptions (Maybe SuggesterFuzzyMatching)

dySearchService :: Lens' DomainStatus (Maybe ServiceEndpoint)

dySearchPartitionCount :: Lens' DomainStatus (Maybe Integer)

dySearchInstanceType :: Lens' DomainStatus (Maybe Text)

dySearchInstanceCount :: Lens' DomainStatus (Maybe Integer)

dyRequiresIndexDocuments :: Lens' DomainStatus Bool

dyProcessing :: Lens' DomainStatus (Maybe Bool)

dyDomainName :: Lens' DomainStatus Text

dyDomainId :: Lens' DomainStatus Text

dyDocService :: Lens' DomainStatus (Maybe ServiceEndpoint)

dyDeleted :: Lens' DomainStatus (Maybe Bool)

dyCreated :: Lens' DomainStatus (Maybe Bool)

dyARN :: Lens' DomainStatus (Maybe Text)

daoSourceFields :: Lens' DoubleArrayOptions (Maybe Text)

daoSearchEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

daoReturnEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

daoFacetEnabled :: Lens' DoubleArrayOptions (Maybe Bool)

daoDefaultValue :: Lens' DoubleArrayOptions (Maybe Double)

ddvSourceField :: Lens' DoubleOptions (Maybe Text)

ddvSortEnabled :: Lens' DoubleOptions (Maybe Bool)

ddvSearchEnabled :: Lens' DoubleOptions (Maybe Bool)

ddvReturnEnabled :: Lens' DoubleOptions (Maybe Bool)

ddvFacetEnabled :: Lens' DoubleOptions (Maybe Bool)

ddvDefaultValue :: Lens' DoubleOptions (Maybe Double)

fExpressionValue :: Lens' Expression Text

fExpressionName :: Lens' Expression Text

esStatus :: Lens' ExpressionStatus OptionStatus

esOptions :: Lens' ExpressionStatus Expression

ifTextOptions :: Lens' IndexField (Maybe TextOptions)

ifTextArrayOptions :: Lens' IndexField (Maybe TextArrayOptions)

ifLiteralOptions :: Lens' IndexField (Maybe LiteralOptions)

ifLiteralArrayOptions :: Lens' IndexField (Maybe LiteralArrayOptions)

ifLatLonOptions :: Lens' IndexField (Maybe LatLonOptions)

ifIntOptions :: Lens' IndexField (Maybe IntOptions)

ifIntArrayOptions :: Lens' IndexField (Maybe IntArrayOptions)

ifIndexFieldType :: Lens' IndexField IndexFieldType

ifIndexFieldName :: Lens' IndexField Text

ifDoubleOptions :: Lens' IndexField (Maybe DoubleOptions)

ifDoubleArrayOptions :: Lens' IndexField (Maybe DoubleArrayOptions)

ifDateOptions :: Lens' IndexField (Maybe DateOptions)

ifDateArrayOptions :: Lens' IndexField (Maybe DateArrayOptions)

ifsStatus :: Lens' IndexFieldStatus OptionStatus

ifsOptions :: Lens' IndexFieldStatus IndexField

iaoSourceFields :: Lens' IntArrayOptions (Maybe Text)

iaoSearchEnabled :: Lens' IntArrayOptions (Maybe Bool)

iaoReturnEnabled :: Lens' IntArrayOptions (Maybe Bool)

iaoFacetEnabled :: Lens' IntArrayOptions (Maybe Bool)

iaoDefaultValue :: Lens' IntArrayOptions (Maybe Integer)

ioSourceField :: Lens' IntOptions (Maybe Text)

ioSortEnabled :: Lens' IntOptions (Maybe Bool)

ioSearchEnabled :: Lens' IntOptions (Maybe Bool)

ioReturnEnabled :: Lens' IntOptions (Maybe Bool)

ioFacetEnabled :: Lens' IntOptions (Maybe Bool)

ioDefaultValue :: Lens' IntOptions (Maybe Integer)

lloSourceField :: Lens' LatLonOptions (Maybe Text)

lloSortEnabled :: Lens' LatLonOptions (Maybe Bool)

lloSearchEnabled :: Lens' LatLonOptions (Maybe Bool)

lloReturnEnabled :: Lens' LatLonOptions (Maybe Bool)

lloFacetEnabled :: Lens' LatLonOptions (Maybe Bool)

lloDefaultValue :: Lens' LatLonOptions (Maybe Text)

laoSourceFields :: Lens' LiteralArrayOptions (Maybe Text)

laoSearchEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

laoReturnEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

laoFacetEnabled :: Lens' LiteralArrayOptions (Maybe Bool)

laoDefaultValue :: Lens' LiteralArrayOptions (Maybe Text)

loSourceField :: Lens' LiteralOptions (Maybe Text)

loSortEnabled :: Lens' LiteralOptions (Maybe Bool)

loSearchEnabled :: Lens' LiteralOptions (Maybe Bool)

loReturnEnabled :: Lens' LiteralOptions (Maybe Bool)

loFacetEnabled :: Lens' LiteralOptions (Maybe Bool)

loDefaultValue :: Lens' LiteralOptions (Maybe Text)

osUpdateVersion :: Lens' OptionStatus (Maybe Integer)

osUpdateDate :: Lens' OptionStatus ISO8601

osState :: Lens' OptionStatus OptionState

osPendingDeletion :: Lens' OptionStatus (Maybe Bool)

osCreationDate :: Lens' OptionStatus ISO8601

spDesiredReplicationCount :: Lens' ScalingParameters (Maybe Integer)

spDesiredPartitionCount :: Lens' ScalingParameters (Maybe Integer)

spDesiredInstanceType :: Lens' ScalingParameters (Maybe PartitionInstanceType)

spsStatus :: Lens' ScalingParametersStatus OptionStatus

spsOptions :: Lens' ScalingParametersStatus ScalingParameters

sSuggesterName :: Lens' Suggester Text

sDocumentSuggesterOptions :: Lens' Suggester DocumentSuggesterOptions

ssStatus :: Lens' SuggesterStatus OptionStatus

ssOptions :: Lens' SuggesterStatus Suggester

taoSourceFields :: Lens' TextArrayOptions (Maybe Text)

taoReturnEnabled :: Lens' TextArrayOptions (Maybe Bool)

taoHighlightEnabled :: Lens' TextArrayOptions (Maybe Bool)

taoDefaultValue :: Lens' TextArrayOptions (Maybe Text)

taoAnalysisScheme :: Lens' TextArrayOptions (Maybe Text)

toSourceField :: Lens' TextOptions (Maybe Text)

toSortEnabled :: Lens' TextOptions (Maybe Bool)

toReturnEnabled :: Lens' TextOptions (Maybe Bool)

toHighlightEnabled :: Lens' TextOptions (Maybe Bool)

toDefaultValue :: Lens' TextOptions (Maybe Text)

toAnalysisScheme :: Lens' TextOptions (Maybe Text)

data UpdateAvailabilityOptions

Constructors

UpdateAvailabilityOptions

Fields

_uaorMultiAZ :: Bool

You expand an existing search domain to a second Availability Zone by setting the Multi-AZ option to true. Similarly, you can turn off the Multi-AZ option to downgrade the domain to a single Availability Zone by setting the Multi-AZ option to false.

_uaorDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

uaorMultiAZ :: Lens' UpdateAvailabilityOptions Bool

uaorDomainName :: Lens' UpdateAvailabilityOptions Text

data UpdateAvailabilityOptionsResponse

Constructors

UpdateAvailabilityOptionsResponse

Fields

_uaosAvailabilityOptions :: Maybe AvailabilityOptionsStatus

The newly-configured availability options. Indicates whether Multi-AZ is enabled for the domain.

uaosAvailabilityOptions :: Iso' UpdateAvailabilityOptionsResponse (Maybe AvailabilityOptionsStatus)

data UpdateScalingParameters

Constructors

UpdateScalingParameters

Fields

_usprDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_usprScalingParameters :: ScalingParameters

The desired instance type and desired number of replicas of each index partition.

usprScalingParameters :: Lens' UpdateScalingParameters ScalingParameters

usprDomainName :: Lens' UpdateScalingParameters Text

data UpdateScalingParametersResponse

Constructors

UpdateScalingParametersResponse

Fields

_uspsScalingParameters :: ScalingParametersStatus

The status and configuration of a search domain's scaling parameters.

uspsScalingParameters :: Iso' UpdateScalingParametersResponse ScalingParametersStatus

data UpdateServiceAccessPolicies

Constructors

UpdateServiceAccessPolicies

Fields

_usaprDomainName :: Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

_usaprAccessPolicies :: Text

The access rules you want to configure. These rules replace any existing rules.

usaprDomainName :: Lens' UpdateServiceAccessPolicies Text

usaprAccessPolicies :: Lens' UpdateServiceAccessPolicies Text

data UpdateServiceAccessPoliciesResponse

Constructors

UpdateServiceAccessPoliciesResponse

Fields

_usapsAccessPolicies :: AccessPoliciesStatus

The access rules configured for the domain.

usapsAccessPolicies :: Iso' UpdateServiceAccessPoliciesResponse AccessPoliciesStatus