Network.AWS.CloudSearch.V2013_01_01.DefineExpression

Description

Configures an Expression for the search domain. Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Request

data DefineExpression

Container for the parameters to the DefineExpression operation. Specifies the name of the domain you want to update and the expression you want to configure.

Request constructor

mkDefineExpression

Arguments

:: Text
deDomainName

-> Expression
deExpression

-> DefineExpression

Smart constructor for the minimum required parameters to construct a valid DefineExpression request.

The fields accessible through corresponding lenses are:

DomainName :: Text

Expression :: Expression

Request lenses

deDomainName :: Lens' DefineExpression Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

deExpression :: Lens' DefineExpression Expression

A named expression that can be evaluated at search time. Can be used for sorting and filtering search results and constructing other expressions.

Response

data DefineExpressionResponse

The result of a DefineExpression request. Contains the status of the newly-configured expression.

Response constructor

mkDefineExpressionResponse

Arguments

:: ExpressionStatus
derExpression

-> DefineExpressionResponse

Smart constructor for the minimum required parameters to construct a valid DefineExpressionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Expression :: ExpressionStatus

Response lenses

derExpression :: Lens' DefineExpressionResponse ExpressionStatus

The value of an Expression and its current status.