Network.AWS.CloudSearch.V2013_01_01.DefineIndexField

Description

Configures an IndexField for the search domain. Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the IndexFieldType. If the field exists, the new configuration replaces the old one. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Request

data DefineIndexField

Container for the parameters to the DefineIndexField operation. Specifies the name of the domain you want to update and the index field configuration.

Request constructor

mkDefineIndexField

Arguments

:: Text
difDomainName

-> IndexField
difIndexField

-> DefineIndexField

Smart constructor for the minimum required parameters to construct a valid DefineIndexField request.

Request lenses

difDomainName :: Lens' DefineIndexField Text

A string that represents the name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

difIndexField :: Lens' DefineIndexField IndexField

The index field and field options you want to configure.

Response

data DefineIndexFieldResponse

The result of a DefineIndexField request. Contains the status of the newly-configured index field.

Response lenses

difrsIndexField :: Lens' DefineIndexFieldResponse IndexFieldStatus

The value of an IndexField and its current status.