Network.AWS.CloudSearch.V2013_01_01.DescribeDomains

Description

Gets information about the search domains owned by this account. Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: q=matchall&amp;q.parser=structured&amp;size=0. For more information, see Getting Information about a Search Domain in the Amazon CloudSearch Developer Guide.

Documentation

describeDomains :: DescribeDomains

Minimum specification for a DescribeDomains request.

data DescribeDomains

Constructors

DescribeDomains

Fields

_ddrDomainNames :: [Text]

The names of the domains you want to include in the response.

ddrDomainNames :: Iso' DescribeDomains [Text]

data DescribeDomainsResponse

Constructors

DescribeDomainsResponse

Fields

_ddsDomainStatusList :: [DomainStatus]

A list that contains the status of each requested domain.

ddsDomainStatusList :: Iso' DescribeDomainsResponse [DomainStatus]